[refactor] 에디터 1차 리팩토링
