+++ Engine/src/Editor/Core/EditorCore.cpp
@@ -3,6 +3,9 @@
 #endif
 
 #include "Editor/Core/EditorCore.h"
+#include "Editor/Core/EditorCommands.h"
+#include "Editor/Core/EditorUIState.h"
+#include "Editor/Core/EditorUndoRedo.h"
 
 #include "Runtime/Rendering/D3D11/ID3D11RenderDevice.h"
 #include "Runtime/Rendering/DeferredRenderSystem.h"
@@ -92,6 +95,21 @@ namespace Alice
 {
 	// 씬 상태 전역 - 여러 네임스페이스에서 공유
 	bool g_SceneDirty = false;
+	bool g_HasCurrentScenePath = false;
+	std::filesystem::path g_CurrentScenePath{};
+	bool g_ShowBuildGameWindow = false;
+	bool g_ShowPvdSettingsWindow = false;
+	bool g_RequestSceneLoad = false;
+	std::filesystem::path g_NextScenePath{};
+	bool g_ShowSceneLoadError = false;
+	std::string g_SceneLoadErrorMsg{};
+	bool g_MaterialEditorOpen = false;
+	std::filesystem::path g_MaterialEditorPath{};
+	MaterialComponent g_MaterialEditorData{};
+	bool g_UICurveEditorOpen = false;
+	std::filesystem::path g_UICurveEditorPath{};
+	UICurveAsset g_UICurveEditorData{};
+	int g_UICurveEditorSelected = -1;
 
 	// ICommand는 이제 EditorCore.h에 정의됨
 
@@ -124,915 +142,12 @@ namespace Alice
 			}
 			return result;
 		}
-		// Build Game 진행 상황 전역 (아래쪽에서 정의됨)
-		extern std::atomic<bool>  g_BuildInProgress;
-		extern std::atomic<float> g_BuildProgress;
-		extern std::atomic<long>  g_BuildExitCode;
-
-		// === ImGuizmo 통합을 위한 어댑터 함수들 ===
-		// 엔진 컨벤션: XMMatrixRotationRollPitchYaw 사용 (런타임과 동일)
-		// rotation.x = pitch, rotation.y = yaw, rotation.z = roll (라디안)
-		inline XMMATRIX BuildRotYPR_Rad(const XMFLOAT3& rotation)
-		{
-			return DirectX::XMMatrixRotationRollPitchYaw(rotation.x, rotation.y, rotation.z);
-		}
-
-		// 로컬 행렬 생성 (S * R * T 순서, row-vector 컨벤션)
-		inline XMMATRIX BuildLocalMatrix(const TransformComponent& transform)
-		{
-			XMMATRIX S = XMMatrixScaling(transform.scale.x, transform.scale.y, transform.scale.z);
-			XMMATRIX R = BuildRotYPR_Rad(transform.rotation);
-			XMMATRIX T = XMMatrixTranslation(transform.position.x, transform.position.y, transform.position.z);
-
-			return S * R * T;
-		}
-
-		// 쿼터니언을 YPR (Euler)로 변환 (쿼터니언 기반 역변환)
-		// PhysicsSystem과 동일한 공식 사용
-		inline XMFLOAT3 QuaternionToYPR_Rad(DirectX::FXMVECTOR q)
-		{
-			DirectX::XMFLOAT4 qq;
-			DirectX::XMStoreFloat4(&qq, q);
-
-			const float x = qq.x, y = qq.y, z = qq.z, w = qq.w;
-
-			float sinp = 2.0f * (w * x - y * z);
-			float pitch = (std::abs(sinp) >= 1.0f)
-				? std::copysign(DirectX::XM_PIDIV2, sinp)
-				: std::asin(sinp);
-
-			float siny_cosp = 2.0f * (w * y + x * z);
-			float cosy_cosp = 1.0f - 2.0f * (x * x + y * y);
-			float yaw = std::atan2(siny_cosp, cosy_cosp);
-
-			float sinr_cosp = 2.0f * (w * z + x * y);
-			float cosr_cosp = 1.0f - 2.0f * (x * x + z * z);
-			float roll = std::atan2(sinr_cosp, cosr_cosp);
-
-			return DirectX::XMFLOAT3(pitch, yaw, roll);
-		}
-
-		// 부모 체인을 따라 올라가며 월드 행렬 계산 (row-vector 컨벤션: World = Local * Parent)
-		// World::ComputeWorldMatrix()를 사용하도록 변경됨
-		// (에디터/런타임 일관성 보장)
-
-		// 로컬 행렬을 TRS로 분해 (쿼터니언 기반 역변환)
-		inline bool DecomposeLocalMatrix(const XMMATRIX& localMatrix, XMFLOAT3& position, XMFLOAT3& rotation, XMFLOAT3& scale)
-		{
-			DirectX::XMVECTOR s, q, t;
-			if (!DirectX::XMMatrixDecompose(&s, &q, &t, localMatrix))
-				return false;
-
-			DirectX::XMStoreFloat3(&position, t);
-			DirectX::XMStoreFloat3(&scale, s);
-			rotation = QuaternionToYPR_Rad(q);
-			return true;
-		}
-
-		// 엔티티 생성 명령
-		struct CreateEntityCommand : ICommand
-		{
-			EntityId entityId;
-			std::string entityType;
-			mutable std::string description; // GetDescription에서 사용
-
-			CreateEntityCommand(EntityId id, const std::string& type)
-				: entityId(id), entityType(type)
-			{
-				description = "Create " + entityType;
-			}
-
-			void Execute(World& world, EntityId& selectedEntity) override
-			{
-				// 엔티티는 이미 생성되어 있음
-				selectedEntity = entityId;
-			}
-
-			void Undo(World& world, EntityId& selectedEntity) override
-			{
-				if (selectedEntity == entityId)
-					selectedEntity = InvalidEntityId;
-				world.DestroyEntity(entityId);
-			}
-
-			const char* GetDescription() const override
-			{
-				return description.c_str();
-			}
-
-			// Create/Destroy는 ID 변경 문제로 Redo 지원 안 함
-			bool SupportsRedo() const override { return false; }
-		};
-
-		// 엔티티 삭제 명령 (씬 파일 직렬화 함수 사용)
-		struct DestroyEntityCommand : ICommand
-		{
-			EntityId entityId;
-			std::string entityName;
-			JsonRttr::json serializedData; // 엔티티 전체를 JSON으로 저장
-			JsonRttr::json childrenData; // 자식 엔티티들의 JSON 데이터 배열
-			mutable std::string description;
-
-			DestroyEntityCommand(EntityId id, const std::string& name, const World& world)
-				: entityId(id), entityName(name)
-			{
-				description = "Delete Entity";
-				// 엔티티 삭제 전에 전체 상태를 JSON으로 저장
-				// SceneFile의 WriteEntity 함수와 동일한 방식 사용
-				JsonRttr::json entityJson;
-				if (WriteEntityToJson(entityJson, world, id))
-				{
-					serializedData = entityJson;
-				}
-
-				// 자식 엔티티들도 저장 (재귀적으로)
-				childrenData = JsonRttr::json::array();
-				SaveChildrenRecursive(world, id, childrenData);
-			}
-
-			// 재귀적으로 자식 엔티티들을 저장하는 헬퍼 함수
-			static void SaveChildrenRecursive(const World& world, EntityId parentId, JsonRttr::json& outArray)
-			{
-				std::vector<EntityId> children = world.GetChildren(parentId);
-				for (EntityId childId : children)
-				{
-					JsonRttr::json childJson;
-					if (WriteEntityToJson(childJson, world, childId))
-					{
-						// 재귀적으로 손자들도 저장 (push_back 전에 완료)
-						JsonRttr::json grandchildrenArray = JsonRttr::json::array();
-						SaveChildrenRecursive(world, childId, grandchildrenArray);
-						if (!grandchildrenArray.empty())
-						{
-							childJson["_children"] = std::move(grandchildrenArray);
-						}
-
-						// 손자 정보를 포함한 childJson을 push
-						outArray.push_back(std::move(childJson));
-					}
-				}
-			}
-
-			// WriteEntity 함수를 복사한 헬퍼 함수 (SceneFile::WriteEntity와 동일한 방식)
-			static bool WriteEntityToJson(JsonRttr::json& outEntity, const World& world, EntityId id)
-			{
-				outEntity = JsonRttr::json::object();
-
-				const std::string name = world.GetEntityName(id);
-				if (!name.empty())
-					outEntity["name"] = name;
-
-				// GUID 저장
-				if (const auto* idComp = world.GetComponent<IDComponent>(id); idComp)
-				{
-					outEntity["guid"] = std::to_string(idComp->guid);
-				}
-
-				// Parent 관계 저장 (GUID 기반)
-				EntityId parentId = world.GetParent(id);
-				if (parentId != InvalidEntityId)
-				{
-					if (const auto* parentIdComp = world.GetComponent<IDComponent>(parentId); parentIdComp)
-					{
-						outEntity["_parentGuid"] = std::to_string(parentIdComp->guid);
-					}
-				}
-
-				if (const auto* transform = world.GetComponent<TransformComponent>(id); transform)
-				{
-					rttr::instance inst = const_cast<TransformComponent&>(*transform);
-					outEntity["Transform"] = JsonRttr::ToJsonObject(inst);
-				}
-
-				if (const auto* scripts = world.GetScripts(id); scripts && !scripts->empty())
-				{
-					JsonRttr::json arr = JsonRttr::json::array();
-					for (const auto& sc : *scripts)
-					{
-						JsonRttr::json s = JsonRttr::json::object();
-						s["name"] = sc.scriptName;
-						s["enabled"] = sc.enabled;
-
-						if (sc.instance)
-						{
-							rttr::instance inst = *sc.instance;
-							const rttr::type t = rttr::type::get_by_name(sc.scriptName);
-							s["props"] = JsonRttr::ToJsonObject(inst, t);
-						}
-
-						arr.push_back(s);
-					}
-					outEntity["Scripts"] = arr;
-				}
-
-				// Material
-				if (const auto* mat = world.GetComponent<MaterialComponent>(id); mat)
-				{
-					MaterialComponent matCopy = *mat;
-					rttr::instance inst = matCopy;
-					outEntity["Material"] = JsonRttr::ToJsonObject(inst);
-				}
-
-				// SkinnedMesh
-				if (const auto* skinned = world.GetComponent<SkinnedMeshComponent>(id); skinned)
-				{
-					SkinnedMeshComponent skinnedCopy = *skinned;
-					rttr::instance inst = skinnedCopy;
-					outEntity["SkinnedMesh"] = JsonRttr::ToJsonObject(inst);
-				}
-
-				// SkinnedAnimation
-				if (const auto* anim = world.GetComponent<SkinnedAnimationComponent>(id); anim)
-				{
-					rttr::instance inst = const_cast<SkinnedAnimationComponent&>(*anim);
-					outEntity["SkinnedAnimation"] = JsonRttr::ToJsonObject(inst);
-				}
-
-				// Socket
-				if (const auto* socketComp = world.GetComponent<SocketComponent>(id); socketComp)
-				{
-					outEntity["Socket"] = SocketSerialization::SocketComponentToJson(*socketComp);
-				}
-
-				// SocketAttachment
-				if (const auto* socketAttach = world.GetComponent<SocketAttachmentComponent>(id); socketAttach)
-				{
-					rttr::instance inst = const_cast<SocketAttachmentComponent&>(*socketAttach);
-					JsonRttr::json obj = JsonRttr::ToJsonObject(inst);
-					obj["ownerGuid"] = std::to_string(socketAttach->ownerGuid);
-					outEntity["SocketAttachment"] = obj;
-				}
-
-				// Camera components
-				if (const auto* cam = world.GetComponent<CameraComponent>(id); cam)
-				{
-					rttr::instance inst = const_cast<CameraComponent&>(*cam);
-					outEntity["Camera"] = JsonRttr::ToJsonObject(inst);
-				}
-				if (const auto* follow = world.GetComponent<CameraFollowComponent>(id); follow)
-				{
-					rttr::instance inst = const_cast<CameraFollowComponent&>(*follow);
-					outEntity["CameraFollow"] = JsonRttr::ToJsonObject(inst);
-				}
-				if (const auto* spring = world.GetComponent<CameraSpringArmComponent>(id); spring)
-				{
-					rttr::instance inst = const_cast<CameraSpringArmComponent&>(*spring);
-					outEntity["CameraSpringArm"] = JsonRttr::ToJsonObject(inst);
-				}
-				if (const auto* lookAt = world.GetComponent<CameraLookAtComponent>(id); lookAt)
-				{
-					rttr::instance inst = const_cast<CameraLookAtComponent&>(*lookAt);
-					outEntity["CameraLookAt"] = JsonRttr::ToJsonObject(inst);
-				}
-				if (const auto* shake = world.GetComponent<CameraShakeComponent>(id); shake)
-				{
-					rttr::instance inst = const_cast<CameraShakeComponent&>(*shake);
-					outEntity["CameraShake"] = JsonRttr::ToJsonObject(inst);
-				}
-				if (const auto* blend = world.GetComponent<CameraBlendComponent>(id); blend)
-				{
-					rttr::instance inst = const_cast<CameraBlendComponent&>(*blend);
-					outEntity["CameraBlend"] = JsonRttr::ToJsonObject(inst);
-				}
-				if (const auto* input = world.GetComponent<CameraInputComponent>(id); input)
-				{
-					rttr::instance inst = const_cast<CameraInputComponent&>(*input);
-					outEntity["CameraInput"] = JsonRttr::ToJsonObject(inst);
-				}
-
-				// Lights
-				if (const auto* pl = world.GetComponent<PointLightComponent>(id); pl)
-				{
-					rttr::instance inst = const_cast<PointLightComponent&>(*pl);
-					outEntity["PointLight"] = JsonRttr::ToJsonObject(inst);
-				}
-				if (const auto* sl = world.GetComponent<SpotLightComponent>(id); sl)
-				{
-					rttr::instance inst = const_cast<SpotLightComponent&>(*sl);
-					outEntity["SpotLight"] = JsonRttr::ToJsonObject(inst);
-				}
-				if (const auto* rl = world.GetComponent<RectLightComponent>(id); rl)
-				{
-					rttr::instance inst = const_cast<RectLightComponent&>(*rl);
-					outEntity["RectLight"] = JsonRttr::ToJsonObject(inst);
-				}
-
-				// Physics Components
-				if (const auto* rigid = world.GetComponent<Phy_RigidBodyComponent>(id); rigid)
-				{
-					rttr::instance inst = const_cast<Phy_RigidBodyComponent&>(*rigid);
-					outEntity["RigidBody"] = JsonRttr::ToJsonObject(inst);
-				}
-				if (const auto* collider = world.GetComponent<Phy_ColliderComponent>(id); collider)
-				{
-					rttr::instance inst = const_cast<Phy_ColliderComponent&>(*collider);
-					outEntity["Collider"] = JsonRttr::ToJsonObject(inst);
-				}
-				if (const auto* meshCollider = world.GetComponent<Phy_MeshColliderComponent>(id); meshCollider)
-				{
-					rttr::instance inst = const_cast<Phy_MeshColliderComponent&>(*meshCollider);
-					outEntity["MeshCollider"] = JsonRttr::ToJsonObject(inst);
-				}
-				if (const auto* cct = world.GetComponent<Phy_CCTComponent>(id); cct)
-				{
-					rttr::instance inst = const_cast<Phy_CCTComponent&>(*cct);
-					outEntity["CharacterController"] = JsonRttr::ToJsonObject(inst);
-				}
-				if (const auto* terrain = world.GetComponent<Phy_TerrainHeightFieldComponent>(id); terrain)
-				{
-					rttr::instance inst = const_cast<Phy_TerrainHeightFieldComponent&>(*terrain);
-					outEntity["TerrainHeightField"] = JsonRttr::ToJsonObject(inst);
-				}
-				if (const auto* joint = world.GetComponent<Phy_JointComponent>(id); joint)
-				{
-					rttr::instance inst = const_cast<Phy_JointComponent&>(*joint);
-					outEntity["Joint"] = JsonRttr::ToJsonObject(inst);
-				}
-
-				return true;
-			}
-
-			// GUID 파싱 헬퍼 (SceneFile와 동일)
-			static std::uint64_t ParseGuid(const JsonRttr::json& j)
-			{
-				if (j.is_string())
-				{
-					try
-					{
-						return std::stoull(j.get<std::string>());
-					}
-					catch (...)
-					{
-						// GUID 생성 함수는 World.cpp에 있으므로 여기서는 0 반환 (나중에 덮어쓰기)
-						return 0;
-					}
-				}
-				else if (j.is_number_unsigned())
-				{
-					return j.get<std::uint64_t>();
-				}
-				return 0;
-			}
-
-			// ApplyEntity 함수를 복사한 헬퍼 함수
-			static bool RestoreEntityFromJson(World& world, const JsonRttr::json& e, EntityId& restoredId)
-			{
-				if (!e.is_object()) return false;
-
-				const EntityId id = world.CreateEntity();
-				restoredId = id;
-
-				// RAII 가드: 복원 실패 시 엔티티가 찌꺼기로 남지 않게 자동 정리
-				struct EntityGuard {
-					World& world;
-					EntityId id;
-					bool committed = false;
-
-					EntityGuard(World& w, EntityId i) : world(w), id(i) {}
-					~EntityGuard() {
-						if (!committed) {
-							world.DestroyEntity(id);
-						}
-					}
-				} guard(world, id);
-
-				const std::string name = e.value("name", std::string{});
-				if (!name.empty())
-					world.SetEntityName(id, name);
-
-				// IDComponent: GUID 복원 (저장된 값으로 덮어쓰기)
-				auto* idComp = world.GetComponent<IDComponent>(id);
-				if (idComp)
-				{
-					if (auto itGuid = e.find("guid"); itGuid != e.end())
-					{
-						auto parsed = ParseGuid(*itGuid);
-						if (parsed != 0) idComp->guid = parsed; // 실패면 덮어쓰지 않기
-					}
-					// 없으면 CreateEntity에서 생성한 GUID 유지
-				}
-
-				// Transform
-				TransformComponent& t = world.AddComponent<TransformComponent>(id);
-				auto itT = e.find("Transform");
-				if (itT != e.end())
-				{
-					rttr::instance inst = t;
-					if (!JsonRttr::FromJsonObject(inst, *itT)) return false;
-					if (itT->is_object() && itT->find("visible") == itT->end())
-					{
-						auto itLegacy = itT->find("renderEnabled");
-						if (itLegacy != itT->end())
-						{
-							if (itLegacy->is_boolean())
-								t.visible = itLegacy->get<bool>();
-							else if (itLegacy->is_number())
-								t.visible = (itLegacy->get<double>() != 0.0);
-						}
-					}
-				}
-
-				// Scripts
-				auto itS = e.find("Scripts");
-				if (itS != e.end() && itS->is_array())
-				{
-					for (const auto& s : *itS)
-					{
-						if (!s.is_object()) continue;
-						const std::string scriptName = s.value("name", std::string{});
-						if (scriptName.empty()) continue;
-
-						ScriptComponent& sc = world.AddScript(id, scriptName);
-						sc.enabled = s.value("enabled", true);
-
-						auto itP = s.find("props");
-						if (itP != s.end() && itP->is_object() && sc.instance)
-						{
-							rttr::instance inst = *sc.instance;
-							const rttr::type t = rttr::type::get_by_name(sc.scriptName);
-							if (!JsonRttr::FromJsonObject(inst, *itP, t)) return false;
-						}
-					}
-				}
-
-				// Material
-				auto itM = e.find("Material");
-				if (itM != e.end() && itM->is_object())
-				{
-					MaterialComponent& mc = world.AddComponent<MaterialComponent>(id, DirectX::XMFLOAT3(0.7f, 0.7f, 0.7f));
-					rttr::instance inst = mc;
-					if (!JsonRttr::FromJsonObject(inst, *itM)) return false;
-				}
-
-				// SkinnedMesh
-				auto itSM = e.find("SkinnedMesh");
-				if (itSM != e.end() && itSM->is_object())
-				{
-					SkinnedMeshComponent tmp;
-					rttr::instance instTmp = tmp;
-					if (!JsonRttr::FromJsonObject(instTmp, *itSM)) return false;
-
-					if (!tmp.meshAssetPath.empty())
-					{
-						SkinnedMeshComponent& sm = world.AddComponent<SkinnedMeshComponent>(id, tmp.meshAssetPath);
-						sm.instanceAssetPath = tmp.instanceAssetPath;
-						static DirectX::XMFLOAT4X4 identityBone = DirectX::XMFLOAT4X4(1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1);
-						sm.boneMatrices = &identityBone;
-						sm.boneCount = 1;
-					}
-				}
-
-				// SkinnedAnimation
-				auto itSA = e.find("SkinnedAnimation");
-				if (itSA != e.end() && itSA->is_object())
-				{
-					SkinnedAnimationComponent& sa = world.AddComponent<SkinnedAnimationComponent>(id);
-					rttr::instance inst = sa;
-					if (!JsonRttr::FromJsonObject(inst, *itSA)) return false;
-				}
-
-				// Socket
-				auto itSocket = e.find("Socket");
-				if (itSocket != e.end() && itSocket->is_object())
-				{
-					SocketComponent& sc = world.AddComponent<SocketComponent>(id);
-					if (!SocketSerialization::JsonToSocketComponent(*itSocket, sc)) return false;
-				}
-
-				// SocketAttachment
-				auto itSAc = e.find("SocketAttachment");
-				if (itSAc != e.end() && itSAc->is_object())
-				{
-					SocketAttachmentComponent& sa = world.AddComponent<SocketAttachmentComponent>(id);
-					if (auto itGuid = itSAc->find("ownerGuid"); itGuid != itSAc->end())
-						sa.ownerGuid = ParseGuid(*itGuid);
-
-					JsonRttr::json copy = *itSAc;
-					copy.erase("ownerGuid");
-					rttr::instance inst = sa;
-					if (!JsonRttr::FromJsonObject(inst, copy)) return false;
-				}
-
-				// Camera components
-				auto itC = e.find("Camera");
-				if (itC != e.end() && itC->is_object())
-				{
-					CameraComponent& cc = world.AddComponent<CameraComponent>(id);
-					rttr::instance inst = cc;
-					if (!JsonRttr::FromJsonObject(inst, *itC)) return false;
-				}
-				auto itCF = e.find("CameraFollow");
-				if (itCF != e.end() && itCF->is_object())
-				{
-					CameraFollowComponent& cf = world.AddComponent<CameraFollowComponent>(id);
-					rttr::instance inst = cf;
-					if (!JsonRttr::FromJsonObject(inst, *itCF)) return false;
-				}
-				auto itSpring = e.find("CameraSpringArm");
-				if (itSpring != e.end() && itSpring->is_object())
-				{
-					CameraSpringArmComponent& sa = world.AddComponent<CameraSpringArmComponent>(id);
-					rttr::instance inst = sa;
-					if (!JsonRttr::FromJsonObject(inst, *itSpring)) return false;
-				}
-				auto itLA = e.find("CameraLookAt");
-				if (itLA != e.end() && itLA->is_object())
-				{
-					CameraLookAtComponent& la = world.AddComponent<CameraLookAtComponent>(id);
-					rttr::instance inst = la;
-					if (!JsonRttr::FromJsonObject(inst, *itLA)) return false;
-				}
-				auto itCS = e.find("CameraShake");
-				if (itCS != e.end() && itCS->is_object())
-				{
-					CameraShakeComponent& cs = world.AddComponent<CameraShakeComponent>(id);
-					rttr::instance inst = cs;
-					if (!JsonRttr::FromJsonObject(inst, *itCS)) return false;
-				}
-				auto itCB = e.find("CameraBlend");
-				if (itCB != e.end() && itCB->is_object())
-				{
-					CameraBlendComponent& cb = world.AddComponent<CameraBlendComponent>(id);
-					rttr::instance inst = cb;
-					if (!JsonRttr::FromJsonObject(inst, *itCB)) return false;
-				}
-				auto itCI = e.find("CameraInput");
-				if (itCI != e.end() && itCI->is_object())
-				{
-					CameraInputComponent& ci = world.AddComponent<CameraInputComponent>(id);
-					rttr::instance inst = ci;
-					if (!JsonRttr::FromJsonObject(inst, *itCI)) return false;
-				}
-
-				// Lights
-				auto itPL = e.find("PointLight");
-				if (itPL != e.end() && itPL->is_object())
-				{
-					PointLightComponent& pl = world.AddComponent<PointLightComponent>(id);
-					rttr::instance inst = pl;
-					if (!JsonRttr::FromJsonObject(inst, *itPL)) return false;
-				}
-				auto itSL = e.find("SpotLight");
-				if (itSL != e.end() && itSL->is_object())
-				{
-					SpotLightComponent& sl = world.AddComponent<SpotLightComponent>(id);
-					rttr::instance inst = sl;
-					if (!JsonRttr::FromJsonObject(inst, *itSL)) return false;
-				}
-				auto itRL = e.find("RectLight");
-				if (itRL != e.end() && itRL->is_object())
-				{
-					RectLightComponent& rl = world.AddComponent<RectLightComponent>(id);
-					rttr::instance inst = rl;
-					if (!JsonRttr::FromJsonObject(inst, *itRL)) return false;
-				}
-
-				// Physics Components
-				auto itRB = e.find("RigidBody");
-				if (itRB != e.end() && itRB->is_object())
-				{
-					Phy_RigidBodyComponent& rb = world.AddComponent<Phy_RigidBodyComponent>(id);
-					rttr::instance inst = rb;
-					if (!JsonRttr::FromJsonObject(inst, *itRB)) return false;
-				}
-				auto itCollider = e.find("Collider");
-				if (itCollider != e.end() && itCollider->is_object())
-				{
-					Phy_ColliderComponent& col = world.AddComponent<Phy_ColliderComponent>(id);
-					rttr::instance inst = col;
-					if (!JsonRttr::FromJsonObject(inst, *itCollider)) return false;
-				}
-				auto itMeshCollider = e.find("MeshCollider");
-				if (itMeshCollider != e.end() && itMeshCollider->is_object())
-				{
-					Phy_MeshColliderComponent& mc = world.AddComponent<Phy_MeshColliderComponent>(id);
-					rttr::instance inst = mc;
-					if (!JsonRttr::FromJsonObject(inst, *itMeshCollider)) return false;
-				}
-				auto itCCT = e.find("CharacterController");
-				if (itCCT != e.end() && itCCT->is_object())
-				{
-					Phy_CCTComponent& cct = world.AddComponent<Phy_CCTComponent>(id);
-					rttr::instance inst = cct;
-					if (!JsonRttr::FromJsonObject(inst, *itCCT)) return false;
-				}
-				auto itTerrain = e.find("TerrainHeightField");
-				if (itTerrain != e.end() && itTerrain->is_object())
-				{
-					Phy_TerrainHeightFieldComponent& terrain = world.AddComponent<Phy_TerrainHeightFieldComponent>(id);
-					rttr::instance inst = terrain;
-					if (!JsonRttr::FromJsonObject(inst, *itTerrain)) return false;
-				}
-				auto itJoint = e.find("Joint");
-				if (itJoint != e.end() && itJoint->is_object())
-				{
-					Phy_JointComponent& joint = world.AddComponent<Phy_JointComponent>(id);
-					rttr::instance inst = joint;
-					if (!JsonRttr::FromJsonObject(inst, *itJoint)) return false;
-				}
-
-				// Parent 관계는 RestoreEntityFromJson 내부에서 처리하지 않음
-				// Undo에서 GUID로 찾아서 연결 (바깥에서 처리)
-
-				// 모든 컴포넌트 복원이 성공했으므로 가드 커밋
-				guard.committed = true;
-				return true;
-			}
-
-			void Execute(World& world, EntityId& selectedEntity) override
-			{
-				if (selectedEntity == entityId)
-					selectedEntity = InvalidEntityId;
-				world.DestroyEntity(entityId);
-			}
-
-			// GUID로 엔티티 찾기 헬퍼
-			static EntityId FindEntityByGuid(World& world, std::uint64_t guid)
-			{
-				for (auto&& [eid, idc] : world.GetComponents<IDComponent>())
-				{
-					if (idc.guid == guid)
-						return eid;
-				}
-				return InvalidEntityId;
-			}
-
-			void Undo(World& world, EntityId& selectedEntity) override
-			{
-				// 엔티티 복원
-				if (!serializedData.is_null())
-				{
-					EntityId restoredId = InvalidEntityId;
-					if (RestoreEntityFromJson(world, serializedData, restoredId))
-					{
-						// 복원된 엔티티를 선택
-						selectedEntity = restoredId;
-
-						// 루트의 외부 부모 복원 (GUID 기반)
-						if (serializedData.contains("_parentGuid"))
-						{
-							std::uint64_t parentGuid = ParseGuid(serializedData["_parentGuid"]);
-							EntityId parent = FindEntityByGuid(world, parentGuid);
-							if (parent != InvalidEntityId)
-							{
-								world.SetParent(restoredId, parent, false);
-							}
-						}
-
-						// 자식 엔티티들도 재귀적으로 복원하고 부모 관계 설정
-						if (childrenData.is_array())
-						{
-							RestoreChildrenRecursive(world, restoredId, childrenData);
-						}
-					}
-				}
-			}
-
-			// 재귀적으로 자식 엔티티들을 복원하는 헬퍼 함수
-			static void RestoreChildrenRecursive(World& world, EntityId parentId, const JsonRttr::json& childrenArray)
-			{
-				if (!childrenArray.is_array()) return;
-
-				for (const auto& childJson : childrenArray)
-				{
-					EntityId restoredChildId = InvalidEntityId;
-					if (RestoreEntityFromJson(world, childJson, restoredChildId))
-					{
-						// 부모 관계 복원 (keepWorld=false: 로드이므로)
-						world.SetParent(restoredChildId, parentId, false);
-
-						// 손자들도 재귀적으로 복원
-						auto it = childJson.find("_children");
-						if (it != childJson.end() && it->is_array())
-						{
-							RestoreChildrenRecursive(world, restoredChildId, *it);
-						}
-					}
-				}
-			}
-
-			const char* GetDescription() const override
-			{
-				return description.c_str();
-			}
-
-			// Create/Destroy는 ID 변경 문제로 Redo 지원 안 함
-			bool SupportsRedo() const override { return false; }
-		};
-
-		// 엔티티 이름 변경 명령
-		struct SetEntityNameCommand : ICommand
-		{
-			EntityId entityId;
-			std::string oldName;
-			std::string newName;
-			mutable std::string description;
-
-			SetEntityNameCommand(EntityId id, const std::string& old, const std::string& nnew)
-				: entityId(id), oldName(old), newName(nnew)
-			{
-				description = "Rename Entity";
-			}
-
-			void Execute(World& world, EntityId& selectedEntity) override
-			{
-				world.SetEntityName(entityId, newName);
-			}
-
-			void Undo(World& world, EntityId& selectedEntity) override
-			{
-				world.SetEntityName(entityId, oldName);
-			}
-
-			const char* GetDescription() const override
-			{
-				return description.c_str();
-			}
-		};
-
-		// Transform 변경 명령
-		struct TransformCommand : ICommand
-		{
-			EntityId entityId;
-			struct TransformData
-			{
-				DirectX::XMFLOAT3 position;
-				DirectX::XMFLOAT3 rotation;
-				DirectX::XMFLOAT3 scale;
-				bool enabled;
-				bool visible;
-			};
-			TransformData oldData;
-			TransformData newData;
-			mutable std::string description;
-
-			TransformCommand(EntityId id, const TransformData& old, const TransformData& nnew)
-				: entityId(id), oldData(old), newData(nnew)
-			{
-				description = "Transform Change";
-			}
-
-			void Execute(World& world, EntityId& selectedEntity) override
-			{
-				if (auto* transform = world.GetComponent<TransformComponent>(entityId))
-				{
-					transform->position = newData.position;
-					transform->rotation = newData.rotation;
-					transform->scale = newData.scale;
-					transform->enabled = newData.enabled;
-					transform->visible = newData.visible;
-					world.MarkTransformDirty(entityId);
-				}
-			}
-
-			void Undo(World& world, EntityId& selectedEntity) override
-			{
-				if (auto* transform = world.GetComponent<TransformComponent>(entityId))
-				{
-					transform->position = oldData.position;
-					transform->rotation = oldData.rotation;
-					transform->scale = oldData.scale;
-					transform->enabled = oldData.enabled;
-					transform->visible = oldData.visible;
-					world.MarkTransformDirty(entityId);
-				}
-			}
-
-			const char* GetDescription() const override
-			{
-				return description.c_str();
-			}
-		};
-
-		// ComponentEditCommand는 이제 EditorCore.h에 정의됨
-
-		// 부모 설정 명령
-		struct SetParentCommand : ICommand
-		{
-			EntityId childId;
-			EntityId oldParent;
-			EntityId newParent;
-			TransformComponent oldLocal;
-			TransformComponent newLocal;
-			bool hasLocalSnapshots;
-			mutable std::string description;
-
-			// 하이라키 드래그용: Transform 스냅샷 포함
-			SetParentCommand(EntityId child, EntityId oldP, EntityId newP,
-				const TransformComponent& oldT, const TransformComponent& newT)
-				: childId(child), oldParent(oldP), newParent(newP),
-				oldLocal(oldT), newLocal(newT), hasLocalSnapshots(true)
-			{
-				description = "Set Parent";
-			}
-
-			// 레거시 호환용: Transform 스냅샷 없음 (keepWorld=false로 동작)
-			SetParentCommand(EntityId child, EntityId oldP, EntityId newP)
-				: childId(child), oldParent(oldP), newParent(newP), hasLocalSnapshots(false)
-			{
-				description = "Set Parent";
-			}
-
-			void Execute(World& world, EntityId& selectedEntity) override
-			{
-				if (hasLocalSnapshots)
-				{
-					// 저장된 로컬 값 사용
-					world.SetParent(childId, newParent, false);
-					if (auto* t = world.GetComponent<TransformComponent>(childId))
-					{
-						*t = newLocal;
-					}
-				}
-				else
-				{
-					// 레거시: keepWorld=false
-					world.SetParent(childId, newParent, false);
-				}
-			}
-
-			void Undo(World& world, EntityId& selectedEntity) override
-			{
-				if (hasLocalSnapshots)
-				{
-					// 저장된 로컬 값 사용
-					world.SetParent(childId, oldParent, false);
-					if (auto* t = world.GetComponent<TransformComponent>(childId))
-					{
-						*t = oldLocal;
-					}
-				}
-				else
-				{
-					// 레거시: keepWorld=false
-					world.SetParent(childId, oldParent, false);
-				}
-			}
-
-			const char* GetDescription() const override
-			{
-				return description.c_str();
-			}
-		};
-
 		// Undo/Redo 스택 관리
 		std::vector<std::unique_ptr<ICommand>> g_UndoStack;
 		std::vector<std::unique_ptr<ICommand>> g_RedoStack;
 		constexpr size_t MAX_UNDO_STACK_SIZE = 50;
 
 		// PushCommand는 EditorCore 클래스의 멤버 함수로 이동됨
-
-		bool ExecuteUndo(World& world, EntityId& selectedEntity)
-		{
-			if (g_UndoStack.empty())
-				return false;
-
-			auto cmd = std::move(g_UndoStack.back());
-			g_UndoStack.pop_back();
-
-			cmd->Undo(world, selectedEntity);
-
-			// Redo 지원하는 커맨드만 Redo 스택에 추가
-			if (cmd->SupportsRedo())
-			{
-				g_RedoStack.push_back(std::move(cmd));
-				if (g_RedoStack.size() > MAX_UNDO_STACK_SIZE)
-				{
-					g_RedoStack.erase(g_RedoStack.begin());
-				}
-			}
-			// else: Redo 불가 커맨드는 버림
-
-			g_SceneDirty = true;
-			return true;
-		}
-
-		bool ExecuteRedo(World& world, EntityId& selectedEntity)
-		{
-			if (g_RedoStack.empty())
-				return false;
-
-			auto cmd = std::move(g_RedoStack.back());
-			g_RedoStack.pop_back();
-
-			cmd->Execute(world, selectedEntity);
-
-			// Undo 스택에 다시 추가
-			g_UndoStack.push_back(std::move(cmd));
-			if (g_UndoStack.size() > MAX_UNDO_STACK_SIZE)
-			{
-				g_UndoStack.erase(g_UndoStack.begin());
-			}
-
-			g_SceneDirty = true;
-			return true;
-		}
-
-		void ClearUndoStack()
-		{
-			g_UndoStack.clear();
-			g_RedoStack.clear();
-		}
 
 		inline bool MaterialInspectorFilter(const std::string& propName)
 		{
@@ -1343,472 +458,58 @@ namespace Alice
 			return true;
 		}
 
-		// 빌드/배포용 간단 파일 유틸 (에러는 로그로 남기고, 실패는 false 반환)
-		bool CopyDirTree(const std::filesystem::path& src, const std::filesystem::path& dst)
-		{
-			namespace fs = std::filesystem;
-			std::error_code ec;
-			if (!fs::exists(src, ec) || ec) return true; // 없는 건 스킵
-			fs::create_directories(dst, ec);
-			if (ec)
-			{
-				ALICE_LOG_ERRORF("BuildGame: create_directories failed. dst=\"%s\" (%s)",
-					dst.string().c_str(), ec.message().c_str());
-				return false;
-			}
-			fs::copy(src, dst, fs::copy_options::recursive | fs::copy_options::overwrite_existing, ec);
-			if (ec)
-			{
-				ALICE_LOG_ERRORF("BuildGame: copy dir failed. \"%s\" -> \"%s\" (%s)",
-					src.string().c_str(), dst.string().c_str(), ec.message().c_str());
-				return false;
-			}
-			return true;
-		}
-
-		bool CopyFileOver(const std::filesystem::path& src, const std::filesystem::path& dst)
-		{
-			namespace fs = std::filesystem;
-			std::error_code ec;
-			fs::create_directories(dst.parent_path(), ec);
-			ec.clear();
-			fs::copy_file(src, dst, fs::copy_options::overwrite_existing, ec);
-			if (ec)
-			{
-				ALICE_LOG_ERRORF("BuildGame: copy file failed. \"%s\" -> \"%s\" (%s)",
-					src.string().c_str(), dst.string().c_str(), ec.message().c_str());
-				return false;
-			}
-			return true;
-		}
-
-		bool MakeCleanDir(const std::filesystem::path& dir)
-		{
-			namespace fs = std::filesystem;
-			std::error_code ec;
-			if (fs::exists(dir, ec))
-			{
-				ec.clear();
-				fs::remove_all(dir, ec);
-			}
-			ec.clear();
-			fs::create_directories(dir, ec);
-			if (ec)
-			{
-				ALICE_LOG_ERRORF("BuildGame: create clean dir failed. \"%s\" (%s)",
-					dir.string().c_str(), ec.message().c_str());
-				return false;
-			}
-			return true;
-		}
-
-		// srcRoot의 모든 파일을 dstCookedRoot/<rel>.alice 로 "암호화 저장"합니다(폴더 구조 유지, 확장자는 .alice로 통일).
-		// - 이미 암호화된 .alice 는 그대로 복사합니다(중복 암호화 방지).
-		// - excludePrefixRel(예: "Resource/")로 시작하는 rel 경로는 스킵할 수 있습니다.
-		bool CookAllIntoCookedRoot(const std::filesystem::path& srcRoot,
-			const std::filesystem::path& dstCookedRoot,
-			const std::string& excludePrefixRel = {})
-		{
-			namespace fs = std::filesystem;
-			std::error_code ec;
-			if (!fs::exists(srcRoot, ec) || ec) return true; // 없는 건 스킵
-			if (!fs::is_directory(srcRoot, ec) || ec) return true;
-
-			// 싱글스레드로 도는 버전임. 오류나면 이걸로 빌드ㄱ
-			//Alice::ResourceManager rm;
-			//if (alreadyEncrypted)
-			//{
-			//	// .alice → .alice 로 그대로 복사 (경로/파일명은 rel 기준으로 새로 배치)
-			//	if (!CopyFileOver(inPath, outPath))
-			//		return false;
-			//}
-			//else
-			//{
-			//	// 디버그: 어떤 파일이 어떤 경로로 cook 되는지 전부 로그로 남깁니다.
-			//	ALICE_LOG_INFO("CookFile: in=\"%s\" -> out=\"%s\"",
-			//		inPath.string().c_str(),
-			//		outPath.string().c_str());
-			//	if (!rm.CookAndSave(inPath, outPath))
-			//	{
-			//		ALICE_LOG_ERRORF("BuildGame: CookAndSave failed. in=\"%s\" out=\"%s\"",
-			//			inPath.string().c_str(), outPath.string().c_str());
-			//		return false;
-			//	}
-			//}
-
-			// 빌드할때 Cook으로 변환할때 쓸 멀티쓰레드 잡임
-			// 모든 작업을 벡터에 수집
-			struct Job
-			{
-				fs::path inPath;
-				fs::path outPath;
-				bool alreadyEncrypted;
-			};
-			std::vector<Job> jobs;
-
-			for (fs::recursive_directory_iterator it(srcRoot, ec), end; it != end; it.increment(ec))
-			{
-				if (ec) { ec.clear(); continue; }
-				if (!it->is_regular_file(ec) || ec) { ec.clear(); continue; }
-
-				const fs::path inPath = it->path();
-				fs::path rel = fs::relative(inPath, srcRoot, ec);
-				if (ec) { ec.clear(); continue; }
-
-				const std::string relStr = rel.generic_string();
-				if (!excludePrefixRel.empty() && relStr.rfind(excludePrefixRel, 0) == 0)
-					continue;
-
-				fs::path outPath = dstCookedRoot / rel;
-				outPath.replace_extension(".alice");
-				const std::string ext = inPath.extension().string();
-				const bool alreadyEncrypted = (_stricmp(ext.c_str(), ".alice") == 0);
-
-				jobs.push_back({ inPath, outPath, alreadyEncrypted });
-			}
-
-			if (jobs.empty()) return true;
-
-			// 2단계: 멀티스레드 병렬 처리
-			std::atomic<size_t> nextIdx = 0;
-			std::atomic<bool> success = true;
-			std::mutex logMutex;
-
-			const size_t numThreads = std::max(1u, std::thread::hardware_concurrency());
-			std::vector<std::thread> workers;
-
-			struct WorkerCtx
-			{
-				std::vector<Job>* jobs{};
-				std::atomic<size_t>* nextIdx{};
-				std::atomic<bool>* success{};
-				std::mutex* logMutex{};
-			};
-
-			struct WorkerProc
-			{
-				static void Run(WorkerCtx ctx)
-				{
-					while (true)
-					{
-						const size_t idx = ctx.nextIdx->fetch_add(1);
-						if (idx >= ctx.jobs->size())
-							break;
-
-						const Job& job = (*ctx.jobs)[idx];
-						bool ok = false;
-
-						if (job.alreadyEncrypted)
-						{
-							ok = CopyFileOver(job.inPath, job.outPath);
-						}
-						else
-						{
-							{
-								std::lock_guard<std::mutex> lock(*ctx.logMutex);
-								ALICE_LOG_INFO("CookFile: in=\"%s\" -> out=\"%s\"",
-									job.inPath.string().c_str(),
-									job.outPath.string().c_str());
-							}
-							ok = Alice::ResourceManager::Get().CookAndSave(job.inPath, job.outPath);
-							if (!ok)
-							{
-								std::lock_guard<std::mutex> lock(*ctx.logMutex);
-								ALICE_LOG_ERRORF("BuildGame: CookAndSave failed. in=\"%s\" out=\"%s\"",
-									job.inPath.string().c_str(), job.outPath.string().c_str());
-							}
-						}
-
-						if (!ok)
-							ctx.success->store(false);
-					}
-				}
-			};
-
-			const WorkerCtx ctx{ &jobs, &nextIdx, &success, &logMutex };
-
-			for (size_t i = 0; i < numThreads; ++i)
-				workers.emplace_back(&WorkerProc::Run, ctx);
-
-			for (auto& t : workers)
-				t.join();
-
-			return success.load();
-		}
-
-		void CopyAllDlls(const std::filesystem::path& fromDir, const std::filesystem::path& toDir)
-		{
-			namespace fs = std::filesystem;
-			std::error_code ec;
-			if (!fs::exists(fromDir, ec) || ec) return;
-			fs::create_directories(toDir, ec);
-			ec.clear();
-
-			// 1단계: 모든 DLL 파일 경로 수집
-			std::vector<fs::path> dllFiles;
-			for (fs::directory_iterator it(fromDir, ec), end; it != end; it.increment(ec))
-			{
-				if (ec) { ec.clear(); continue; }
-				if (!it->is_regular_file(ec) || ec) { ec.clear(); continue; }
-				const fs::path p = it->path();
-				if (p.extension() == ".dll")
-				{
-					// 이 부분에서 dllFiles을 푸시해서 멀티스레드 도는  건데, 
-					// 만약 오류가 생긴다면 바로 CopyFileOver로 여기서 싱글스레드로 할것.
-					//CopyFileOver(p, toDir / p.filename());
-					dllFiles.push_back(p);
-				}
-			}
-
-			if (dllFiles.empty()) return;
-
-			// 2단계: 멀티스레드 병렬 복사
-			std::atomic<size_t> nextIdx = 0;
-			const size_t numThreads = std::max(1u, std::thread::hardware_concurrency());
-			std::vector<std::thread> workers;
-
-			struct CopyCtx
-			{
-				std::vector<fs::path>* dllFiles{};
-				std::atomic<size_t>* nextIdx{};
-				fs::path toDir{};
-			};
-
-			struct CopyProc
-			{
-				static void Run(CopyCtx ctx)
-				{
-					while (true)
-					{
-						const size_t idx = ctx.nextIdx->fetch_add(1);
-						if (idx >= ctx.dllFiles->size())
-							break;
-						CopyFileOver((*ctx.dllFiles)[idx], ctx.toDir / (*ctx.dllFiles)[idx].filename());
-					}
-				}
-			};
-
-			const CopyCtx ctx{ &dllFiles, &nextIdx, toDir };
-
-			for (size_t i = 0; i < numThreads; ++i)
-				workers.emplace_back(&CopyProc::Run, ctx);
-
-			for (auto& t : workers)
-				t.join();
-		}
-
-		struct BuildGameTaskArgs
-		{
-			std::filesystem::path projectRoot;
-			std::filesystem::path cfgPath;
-			std::string           exportPathStr;
-		};
-
-		struct BuildGameTask
-		{
-			static void Run(BuildGameTaskArgs args)
-			{
-				namespace fs2 = std::filesystem;
-
-#ifdef _DEBUG
-				const std::wstring cmd = L"cmake --build build --config Debug --target AlicePlayer";
-				const fs2::path releaseBinDir = args.projectRoot / "build/bin/Debug";
-#else
-				const std::wstring cmd = L"cmake --build build --config Release --target AlicePlayer";
-				const fs2::path releaseBinDir = args.projectRoot / "build/bin/Release";
-#endif
-
-				STARTUPINFOW        si{};
-				PROCESS_INFORMATION pi{};
-				si.cb = sizeof(si);
-				si.dwFlags = STARTF_USESHOWWINDOW;
-				si.wShowWindow = SW_HIDE;
-
-				// CreateProcessW는 커맨드라인 버퍼를 수정할 수 있어야 하므로 writable buffer 사용
-				std::vector<wchar_t> cmdBuffer(cmd.begin(), cmd.end());
-				cmdBuffer.push_back(L'\0');
-
-				BOOL ok = CreateProcessW(
-					nullptr,
-					cmdBuffer.data(),  // writable buffer
-					nullptr,
-					nullptr,
-					FALSE,
-					CREATE_NO_WINDOW,
-					nullptr,
-					args.projectRoot.wstring().c_str(),
-					&si,
-					&pi);
-
-				if (!ok)
-				{
-					ALICE_LOG_ERRORF("Build Game: failed to start CMake process.");
-					g_BuildInProgress.store(false);
-					g_BuildProgress.store(0.0f);
-					g_BuildExitCode.store(1);
-					return;
-				}
-
-				ScopedHandle hProcess(pi.hProcess);
-				ScopedHandle hThread(pi.hThread);
-
-				float p = 0.0f;
-				for (;;)
-				{
-					DWORD wait = WaitForSingleObject(hProcess.h, 50);
-					if (wait == WAIT_TIMEOUT)
-					{
-						p += 0.005f;
-						if (p > 0.9f) p = 0.9f;
-						g_BuildProgress.store(p);
-						continue;
-					}
-					break;
-				}
-
-				DWORD exitCode = 0;
-				GetExitCodeProcess(hProcess.h, &exitCode);
-				ALICE_LOG_INFO("Build Game: CMake build finished with exitCode=%lu",
-					static_cast<unsigned long>(exitCode));
-
-				if (exitCode != 0)
-				{
-					g_BuildProgress.store(1.0f);
-					g_BuildExitCode.store(static_cast<long>(exitCode));
-					g_BuildInProgress.store(false);
-					return;
-				}
-
-				// (1) Metas: Assets를 청크로 패킹 (폴더구조 숨김, 256KB)
-				const fs2::path stageMetas = releaseBinDir / "Metas";
-				if (!MakeCleanDir(stageMetas))
-				{
-					g_BuildExitCode.store(2);
-					g_BuildInProgress.store(false);
-					return;
-				}
-				{
-					if (!Alice::ResourceManager::Get().CookResourceToChunkStore(args.projectRoot / "Assets", stageMetas, 256 * 1024))
-					{
-						ALICE_LOG_ERRORF("Build Game: failed to cook Assets -> Metas/Chunks.");
-						g_BuildExitCode.store(3);
-						g_BuildInProgress.store(false);
-						return;
-					}
-				}
-
-				// (2) Cooked: 항상 새로 생성 + 전부 .alice 암호화
-				const fs2::path stageCooked = releaseBinDir / "Cooked";
-				if (!MakeCleanDir(stageCooked))
-				{
-					g_BuildExitCode.store(4);
-					g_BuildInProgress.store(false);
-					return;
-				}
-				if (!CookAllIntoCookedRoot(args.projectRoot / "Cooked", stageCooked, "Resource/"))
-				{
-					g_BuildExitCode.store(5);
-					g_BuildInProgress.store(false);
-					return;
-				}
-
-				// (3) Resource: 원본 폴더를 넣지 않고 Cooked/Chunks로 패킹
-				{
-					if (!Alice::ResourceManager::Get().CookResourceToChunkStore(args.projectRoot / "Resource", stageCooked))
-					{
-						ALICE_LOG_ERRORF("Build Game: failed to cook Resource -> Cooked/Chunks (stage).");
-						g_BuildExitCode.store(6);
-						g_BuildInProgress.store(false);
-						return;
-					}
-				}
-
-				// (4) BuildSettings 복사 (exe 옆)
-				if (!CopyFileOver(args.cfgPath, releaseBinDir / "BuildSettings.json"))
-				{
-					g_BuildExitCode.store(7);
-					g_BuildInProgress.store(false);
-					return;
-				}
-
-				// (5) Export: Bin 아래로 정리 (exe/dll/buildsettings/cooked/metas)
-				fs2::path exportRoot = args.exportPathStr;
-				if (!exportRoot.is_absolute())
-					exportRoot = args.projectRoot / exportRoot;
-
-				const fs2::path exportBin = exportRoot / "Bin";
-				if (!MakeCleanDir(exportBin))
-				{
-					g_BuildExitCode.store(8);
-					g_BuildInProgress.store(false);
-					return;
-				}
-
-				if (!CopyFileOver(releaseBinDir / "AlicePlayer.exe", exportBin / "AlicePlayer.exe"))
-				{
-					g_BuildExitCode.store(9);
-					g_BuildInProgress.store(false);
-					return;
-				}
-
-				CopyAllDlls(releaseBinDir, exportBin);
-
-				if (!CopyDirTree(releaseBinDir / "Cooked", exportBin / "Cooked") ||
-					!CopyDirTree(releaseBinDir / "Metas", exportBin / "Metas"))
-				{
-					g_BuildExitCode.store(10);
-					g_BuildInProgress.store(false);
-					return;
-				}
-
-				if (!CopyFileOver(releaseBinDir / "BuildSettings.json", exportBin / "BuildSettings.json"))
-				{
-					g_BuildExitCode.store(11);
-					g_BuildInProgress.store(false);
-					return;
-				}
-
-				ALICE_LOG_INFO("Build Game: exported to \"%s\" (run: Bin/AlicePlayer.exe)",
-					exportRoot.string().c_str());
-
-				g_BuildProgress.store(1.0f);
-				g_BuildExitCode.store(static_cast<long>(exitCode));
-				g_BuildInProgress.store(false);
-			}
-		};
 	}
 
-	namespace
+	bool ExecuteUndo(World& world, EntityId& selectedEntity)
 	{
-		// 현재 씬이 수정되었는지 여부 (저장 필요 여부) - Alice 네임스페이스 레벨에서 정의됨
-		bool                     g_HasCurrentScenePath = false;
-		std::filesystem::path    g_CurrentScenePath;
+		if (g_UndoStack.empty())
+			return false;
 
-		// 간단한 게임 빌드 UI 상태
-		bool                     g_ShowBuildGameWindow = false;
-		bool                     g_ShowPvdSettingsWindow = false;
+		auto cmd = std::move(g_UndoStack.back());
+		g_UndoStack.pop_back();
 
-		// Build Game 진행 상황 (간단한 멀티스레드 + atomic 사용)
-		std::atomic<bool>        g_BuildInProgress{ false };
-		std::atomic<float>       g_BuildProgress{ 0.0f };   // 0.0 ~ 1.0
-		std::atomic<long>        g_BuildExitCode{ -1 };     // -1: 아직 없음
+		cmd->Undo(world, selectedEntity);
 
-		// 다른 씬을 로드하기 위해 대기 중인 경로
-		bool                     g_RequestSceneLoad = false;
-		std::filesystem::path    g_NextScenePath;
-		bool                     g_ShowSceneLoadError = false;
-		std::string              g_SceneLoadErrorMsg;
+		// Redo 지원하는 커맨드만 Redo 스택에 추가
+		if (cmd->SupportsRedo())
+		{
+			g_RedoStack.push_back(std::move(cmd));
+			if (g_RedoStack.size() > MAX_UNDO_STACK_SIZE)
+			{
+				g_RedoStack.erase(g_RedoStack.begin());
+			}
+		}
+		// else: Redo 불가 커맨드는 버림
 
-		// 단일 머티리얼 에셋 편집기 상태
-		bool                     g_MaterialEditorOpen = false;
-		std::filesystem::path    g_MaterialEditorPath;
-		MaterialComponent        g_MaterialEditorData;
+		g_SceneDirty = true;
+		return true;
+	}
 
-		// UI 커브 에셋 편집기 상태
-		bool                     g_UICurveEditorOpen = false;
-		std::filesystem::path    g_UICurveEditorPath;
-		UICurveAsset             g_UICurveEditorData;
-		int                      g_UICurveEditorSelected = -1;
+	bool ExecuteRedo(World& world, EntityId& selectedEntity)
+	{
+		if (g_RedoStack.empty())
+			return false;
+
+		auto cmd = std::move(g_RedoStack.back());
+		g_RedoStack.pop_back();
+
+		cmd->Execute(world, selectedEntity);
+
+		// Undo 스택에 다시 추가
+		g_UndoStack.push_back(std::move(cmd));
+		if (g_UndoStack.size() > MAX_UNDO_STACK_SIZE)
+		{
+			g_UndoStack.erase(g_UndoStack.begin());
+		}
+
+		g_SceneDirty = true;
+		return true;
+	}
+
+	void ClearUndoStack()
+	{
+		g_UndoStack.clear();
+		g_RedoStack.clear();
 	}
 
 	EditorCore::~EditorCore()
@@ -1939,6 +640,27 @@ namespace Alice
 		ImGui_ImplDX11_RenderDrawData(ImGui::GetDrawData());
 	}
 
+	void EditorCore::HandleGlobalUndoRedo(World& world, EntityId& selectedEntity, bool isPlaying)
+	{
+		ImGuiIO& io = ImGui::GetIO();
+		const bool isTextInputActive = io.WantTextInput || ImGui::IsAnyItemActive();
+
+		if (m_inputSystem && !isTextInputActive && !isPlaying)
+		{
+			const bool ctrlDown = m_inputSystem->IsKeyDown(Keyboard::Keys::LeftControl) ||
+				m_inputSystem->IsKeyDown(Keyboard::Keys::RightControl);
+
+			if (ctrlDown && m_inputSystem->IsKeyPressed(Keyboard::Keys::Z))
+			{
+				ExecuteUndo(world, selectedEntity);
+			}
+			else if (ctrlDown && m_inputSystem->IsKeyPressed(Keyboard::Keys::Y))
+			{
+				ExecuteRedo(world, selectedEntity);
+			}
+		}
+	}
+
 	void EditorCore::DrawEditorUI(World& world,
 		Camera& camera,
 		ForwardRenderSystem& forward,
@@ -1962,7 +684,7 @@ namespace Alice
 		deferred.SetDefaultPostProcessSettings(m_defaultPostProcessSettings);
 		// ForwardRenderSystem에도 동일한 함수가 필요하면 추가
 		// forward.SetDefaultPostProcessSettings(m_defaultPostProcessSettings);
-		
+
 		// SceneManager에서 현재 씬 파일 경로를 조회하여 g_CurrentScenePath 업데이트
 		if (sceneManager)
 		{
@@ -1973,3140 +695,25 @@ namespace Alice
 				g_HasCurrentScenePath = true;
 			}
 		}
-		// === Undo 키 입력 처리 (전역) ===
-		ImGuiIO& io = ImGui::GetIO();
-		const bool isTextInputActive = io.WantTextInput || ImGui::IsAnyItemActive();
 
-		// Gizmo 및 스냅 관련 변수 (뷰창과 인스펙터에서 공유)
-		static ImGuizmo::OPERATION gizmoOp = ImGuizmo::TRANSLATE;
-		static ImGuizmo::MODE gizmoMode = ImGuizmo::WORLD; // 기본값: WORLD 모드
+		HandleGlobalUndoRedo(world, selectedEntity, isPlaying);
+		SetupDockSpaceAndDefaultLayout();
+		DrawMainMenuBar(world, deltaTime, fps, isPlaying, selectedEntity, useForwardRendering, isDebugDraw);
 
-		// 스냅 모드 enum
-		enum class SnapMode {
-			None = 0,
-			Increment = 1,
-			Object = 2
-		};
+		DrawPvdSettingsWindow(pvdEnabled, pvdHost, pvdPort);
+		DrawBuildGameWindow();
 
-		// 오브젝트 스냅 타입 enum (Blender 스타일)
-		enum class ObjectSnapType {
-			Center = 0,  // 중심점 (Transform position)
-			Vertex = 1,  // 버텍스
-			Edge = 2,    // 엣지
-			Face = 3     // 면
-		};
+		DrawHierarchyWindow(world, selectedEntity);
+		DrawInspectorWindow(world, selectedEntity);
+		DrawProjectWindow(world, selectedEntity);
+		DrawGameViewportWindow(world, camera, forward, deferred, selectedEntity, picker, cameraMoveSpeed, useForwardRendering, isPlaying, shadingMode, useFillLight);
+		DrawCameraWindow(world, camera, cameraMoveSpeed, selectedEntity);
+		DrawLightingWindow(world, forward, deferred, shadingMode, useFillLight, useForwardRendering);
 
-		static SnapMode snapMode = SnapMode::None;
-		static ObjectSnapType objectSnapType = ObjectSnapType::Center;
-		static bool gizmoSnap = false; // 레거시 호환성 (Increment 모드와 동일)
-		static XMFLOAT3 snapTranslation = XMFLOAT3(1.0f, 1.0f, 1.0f);
-		static float snapRotation = 15.0f; // degrees
-		static float snapScale = 1.0f;
-		static float objectSnapDistance = 0.5f; // 오브젝트 스냅 거리
+		DrawMaterialAssetEditorWindow(world);
+		DrawUICurveAssetEditorWindow();
 
-		if (m_inputSystem && !isTextInputActive && !isPlaying)
-		{
-			using namespace DirectX;
-			const bool ctrlDown = m_inputSystem->IsKeyDown(Keyboard::Keys::LeftControl) ||
-				m_inputSystem->IsKeyDown(Keyboard::Keys::RightControl);
-
-			if (ctrlDown && m_inputSystem->IsKeyPressed(Keyboard::Keys::Z))
-			{
-				ExecuteUndo(world, selectedEntity);
-			}
-			else if (ctrlDown && m_inputSystem->IsKeyPressed(Keyboard::Keys::Y))
-			{
-				ExecuteRedo(world, selectedEntity);
-			}
-		}
-
-		// 메인 뷰포트 전체를 도킹 스페이스로 사용합니다.
-		ImGuiViewport* viewport = ImGui::GetMainViewport();
-		ImGuiID dockspaceId = ImGui::DockSpaceOverViewport(viewport->ID, viewport);
-
-		// 첫 프레임에만 기본 도킹 레이아웃을 구성합니다.
-		static bool s_dockInitialized = false;
-		if (!s_dockInitialized)
-		{
-			s_dockInitialized = true;
-
-			ImGui::DockBuilderRemoveNode(dockspaceId);
-			ImGui::DockBuilderAddNode(dockspaceId, ImGuiDockNodeFlags_DockSpace);
-			ImGui::DockBuilderSetNodeSize(dockspaceId, viewport->Size);
-
-			ImGuiID dockMain = dockspaceId;
-			ImGuiID dockLeft = 0;
-			ImGuiID dockRight = 0;
-			ImGuiID dockCenter = 0;
-			ImGuiID dockRightCol = 0;
-
-			ImGui::DockBuilderSplitNode(dockMain, ImGuiDir_Left, 0.22f, &dockLeft, &dockRight);
-			ImGui::DockBuilderSplitNode(dockRight, ImGuiDir_Right, 0.26f, &dockRightCol, &dockCenter);
-
-			ImGuiID dockCenterTop = 0;
-			ImGuiID dockCenterBottom = 0;
-			ImGui::DockBuilderSplitNode(dockCenter, ImGuiDir_Down, 0.30f, &dockCenterBottom, &dockCenterTop);
-
-			ImGuiID dockLeftTop = 0;
-			ImGuiID dockLeftBottom = 0;
-			ImGui::DockBuilderSplitNode(dockLeft, ImGuiDir_Down, 0.55f, &dockLeftBottom, &dockLeftTop);
-
-			ImGuiID dockRightTop = 0;
-			ImGuiID dockRightBottom = 0;
-			ImGui::DockBuilderSplitNode(dockRightCol, ImGuiDir_Down, 0.55f, &dockRightBottom, &dockRightTop);
-
-			ImGui::DockBuilderDockWindow("Hierarchy", dockLeftTop);
-			ImGui::DockBuilderDockWindow("Project", dockLeftBottom);
-			ImGui::DockBuilderDockWindow("Game", dockCenterTop);
-			ImGui::DockBuilderDockWindow("Camera", dockCenterBottom);
-			ImGui::DockBuilderDockWindow("Inspector", dockRightTop);
-			ImGui::DockBuilderDockWindow("Lighting", dockRightBottom);
-
-			ImGui::DockBuilderFinish(dockspaceId);
-		}
-
-		// === Toolbar ===
-		if (ImGui::BeginMainMenuBar())
-		{
-			ImGui::Text("AliceRenderer");
-			ImGui::Separator();
-
-			// Play / Stop 토글 버튼
-			if (!isPlaying)
-			{
-				if (ImGui::Button("Play"))
-				{
-					// 플레이 전에 스크립트 리로드 실행
-					ALICE_LOG_INFO("Play button pressed: Starting script reload...");
-					bool reloadSuccess = ReloadScripts_FromButton(world);
-					if (!reloadSuccess)
-					{
-						// 스크립트 리로드 실패 시 경고 표시 및 게임 실행 중단
-						ALICE_LOG_ERRORF("Play button: Script reload failed. Game will NOT start.");
-						ImGui::OpenPopup("ScriptReloadFailed");
-						// isPlaying은 설정하지 않음 (게임 실행 안 함)
-					}
-					else
-					{
-						ALICE_LOG_INFO("Play button: Script reload succeeded. Starting game...");
-						isPlaying = true;
-					}
-				}
-			}
-			else
-			{
-				if (ImGui::Button("Stop"))
-				{
-					isPlaying = false;
-				}
-			}
-
-			// 스크립트 리로드 실패 경고 팝업
-			if (ImGui::BeginPopupModal("ScriptReloadFailed", nullptr, ImGuiWindowFlags_AlwaysAutoResize))
-			{
-				ImGui::Text("Script Reload Failed!");
-				ImGui::Separator();
-				ImGui::Text("Failed to reload scripts before starting the game.");
-				ImGui::Text("Please check the console for error details.");
-				ImGui::Text("The game will not start until scripts are reloaded successfully.");
-				ImGui::Separator();
-				if (ImGui::Button("OK", ImVec2(120, 0)))
-				{
-					ImGui::CloseCurrentPopup();
-				}
-				ImGui::EndPopup();
-			}
-
-			// 오브젝트 생성 메뉴 버튼
-			if (ImGui::Button("Create"))
-			{
-				ImGui::OpenPopup("CreateObjectPopup");
-			}
-			if (ImGui::BeginPopup("CreateObjectPopup"))
-			{
-				if (ImGui::MenuItem("Empty"))
-				{
-					EntityId e = world.CreateEmpty();
-					PushCommand(std::make_unique<CreateEntityCommand>(e, "Empty"));
-					selectedEntity = e;
-					g_SceneDirty = true;
-					ImGui::CloseCurrentPopup();
-				}
-				// FBX Primitives 메뉴
-				if (ImGui::BeginMenu("FBX Primitives"))
-				{
-					// 프리미티브 폴더 스캔해서 자동으로 메뉴 채우기
-					static std::vector<std::filesystem::path> cached;
-					static bool cachedOnce = false;
-
-					if (!cachedOnce)
-					{
-						cachedOnce = true;
-
-						auto dirAbs = ResourceManager::Get().Resolve("Assets/Fbx");
-						if (std::filesystem::exists(dirAbs))
-						{
-							for (auto& it : std::filesystem::directory_iterator(dirAbs))
-							{
-								if (!it.is_regular_file()) continue;
-								auto p = it.path();
-								auto ext = p.extension().string();
-								std::transform(ext.begin(), ext.end(), ext.begin(),
-									[](unsigned char c) { return static_cast<char>(std::tolower(c)); });
-								if (ext == ".fbxasset")
-									cached.push_back(p);
-							}
-							std::sort(cached.begin(), cached.end());
-						}
-					}
-
-					// 고정 프리미티브 목록 (폴더에 없어도 표시)
-					struct Prim { const char* label; const char* path; };
-					static Prim prims[] = {
-						{"IcoSphere", "../Assets/Fbx/IcoSphere.fbxasset"},
-						{"Torus",     "../Assets/Fbx/Torus.fbxasset"},
-						{"Monkey",    "../Assets/Fbx/Monkey.fbxasset"},
-						//{"Box",       "../Assets/Fbx/Box.fbxasset"},
-						{"Cube(FBX)", "../Assets/Fbx/Cube.fbxasset"},
-						{"Sphere(FBX)", "../Assets/Fbx/Sphere.fbxasset"},
-						{"Quad(FBX)", "../Assets/Fbx/Quad.fbxasset"},
-						{"Corn(FBX)", "../Assets/Fbx/Corn.fbxasset"}
-					};
-
-					// 고정 목록 표시
-					for (auto& p : prims)
-					{
-						if (ImGui::MenuItem(p.label))
-						{
-							EntityId e = InstantiateFbxAssetToWorld(world, p.path, p.label);
-							if (e != InvalidEntityId)
-							{
-								PushCommand(std::make_unique<CreateEntityCommand>(e, p.label));
-								selectedEntity = e;
-							}
-							ImGui::CloseCurrentPopup();
-						}
-					}
-
-					// 폴더에서 스캔한 추가 FBX들 표시
-					if (!cached.empty())
-					{
-						ImGui::Separator();
-						for (auto& abs : cached)
-						{
-							std::string label = abs.stem().string();
-
-							// 이미 고정 목록에 있는 건 스킵
-							bool skip = false; 
-							for (auto& p : prims)
-							{
-								if (label == p.label || label == "Cube" && std::string(p.label) == "Cube(FBX)")
-								{
-									skip = true;
-									break;
-								}
-							}
-							if (skip) continue;
-
-							if (ImGui::MenuItem(label.c_str()))
-							{
-								EntityId e = InstantiateFbxAssetToWorld(world, abs, label);
-								if (e != InvalidEntityId)
-								{
-									PushCommand(std::make_unique<CreateEntityCommand>(e, label));
-									selectedEntity = e;
-								}
-								ImGui::CloseCurrentPopup();
-							}
-						}
-					}
-
-					ImGui::EndMenu();
-				}
-
-				if (ImGui::MenuItem("Camera"))
-				{
-					EntityId e = world.CreateCamera();
-					PushCommand(std::make_unique<CreateEntityCommand>(e, "Camera"));
-					selectedEntity = e;
-					g_SceneDirty = true;
-					ImGui::CloseCurrentPopup();
-				}
-				if (ImGui::MenuItem("Point Light"))
-				{
-					EntityId e = world.CreatePointLight();
-					PushCommand(std::make_unique<CreateEntityCommand>(e, "Point Light"));
-					selectedEntity = e;
-					g_SceneDirty = true;
-					ImGui::CloseCurrentPopup();
-				}
-				if (ImGui::MenuItem("Spot Light"))
-				{
-					EntityId e = world.CreateSpotLight();
-					PushCommand(std::make_unique<CreateEntityCommand>(e, "Spot Light"));
-					selectedEntity = e;
-					g_SceneDirty = true;
-					ImGui::CloseCurrentPopup();
-				}
-				if (ImGui::MenuItem("Rect Light"))
-				{
-					EntityId e = world.CreateRectLight();
-					PushCommand(std::make_unique<CreateEntityCommand>(e, "Rect Light"));
-                    selectedEntity = e;
-                    g_SceneDirty = true;
-                    ImGui::CloseCurrentPopup();
-                }
-				if (ImGui::BeginMenu("AliceUI"))
-				{
-					if (ImGui::MenuItem("Screen Image"))
-					{
-						EntityId e = CreateAliceUIImage(world);
-						if (e != InvalidEntityId)
-						{
-							PushCommand(std::make_unique<CreateEntityCommand>(e, "UI Image"));
-							selectedEntity = e;
-							g_SceneDirty = true;
-						}
-						ImGui::CloseCurrentPopup();
-					}
-					if (ImGui::MenuItem("Screen Text"))
-					{
-						EntityId e = CreateAliceUIText(world);
-						if (e != InvalidEntityId)
-						{
-							PushCommand(std::make_unique<CreateEntityCommand>(e, "UI Text"));
-							selectedEntity = e;
-							g_SceneDirty = true;
-						}
-						ImGui::CloseCurrentPopup();
-					}
-					if (ImGui::MenuItem("Screen Button"))
-					{
-						EntityId e = CreateAliceUIButton(world);
-						if (e != InvalidEntityId)
-						{
-							PushCommand(std::make_unique<CreateEntityCommand>(e, "UI Button"));
-							selectedEntity = e;
-							g_SceneDirty = true;
-						}
-						ImGui::CloseCurrentPopup();
-					}
-					if (ImGui::MenuItem("Screen Gauge"))
-					{
-						EntityId e = CreateAliceUIGauge(world);
-						if (e != InvalidEntityId)
-						{
-							PushCommand(std::make_unique<CreateEntityCommand>(e, "UI Gauge"));
-							selectedEntity = e;
-							g_SceneDirty = true;
-						}
-						ImGui::CloseCurrentPopup();
-					}
-					if (ImGui::MenuItem("World Image"))
-					{
-						EntityId e = CreateAliceUIWorldImage(world);
-						if (e != InvalidEntityId)
-						{
-							PushCommand(std::make_unique<CreateEntityCommand>(e, "World UI Image"));
-							selectedEntity = e;
-							g_SceneDirty = true;
-						}
-						ImGui::CloseCurrentPopup();
-					}
-					ImGui::EndMenu();
-				}
-                ImGui::EndPopup();
-            }
-
-			ImGui::Separator();
-
-			// 스크립트 핫 리로드 버튼 (C++ 스크립트 DLL 재빌드 + 재로드)
-			if (ImGui::Button("Reload Scripts"))
-			{
-				// ImGui Begin/End 짝을 깨지 않기 위해,
-				// 실제 빌드/복사/리로드 로직은 별도 헬퍼 함수에서 처리합니다.
-				ReloadScripts_FromButton(world);
-				m_scriptBuilded = true;
-			}
-
-			ImGui::Separator();
-			// FBX 임포트 버튼
-			if (ImGui::Button("Load FBX"))
-			{
-				wchar_t fileBuffer[MAX_PATH] = {};
-				OPENFILENAMEW ofn{};
-				ofn.lStructSize = sizeof(ofn);
-				ofn.hwndOwner = m_hwnd;
-				ofn.lpstrFilter = L"FBX Files\0*.fbx\0All Files\0*.*\0";
-				ofn.lpstrFile = fileBuffer;
-				ofn.nMaxFile = MAX_PATH;
-				ofn.Flags = OFN_EXPLORER | OFN_FILEMUSTEXIST | OFN_PATHMUSTEXIST | OFN_NOCHANGEDIR;
-
-				if (GetOpenFileNameW(&ofn))
-				{
-					if (m_renderDevice)
-					{
-						std::filesystem::path fbxPath = fileBuffer;
-
-						wchar_t exePathW[MAX_PATH] = {};
-						GetModuleFileNameW(nullptr, exePathW, MAX_PATH);
-						std::filesystem::path exePath = exePathW;
-						std::filesystem::path exeDir = exePath.parent_path();
-						std::filesystem::path projectRoot = exeDir.parent_path().parent_path().parent_path(); // build/bin/Debug → 프로젝트 루트
-
-						fbxPath = std::filesystem::relative(fbxPath, projectRoot);
-
-						// 간단한 FBX 임포트 옵션
-						FbxImportOptions opt{};
-						FbxImporter importer(ResourceManager::Get(), m_skinnedRegistry);
-
-						auto* d3dDevice = m_renderDevice->GetDevice();
-						FbxImportResult result = importer.Import(d3dDevice, fbxPath, opt);
-
-						// 1) 인스턴스 에셋(.fbxasset)이 생성되었으면, 프로젝트 뷰에서 활용할 수 있습니다.
-						// 2) 월드에 기본 인스턴스 하나를 바로 생성해 줍니다. (언리얼의 "씬에 배치" 느낌)
-						if (!result.meshAssetPath.empty())
-						{
-							EntityId e = world.CreateEntity();
-							TransformComponent& t = world.AddComponent<TransformComponent>(e);
-							t.position = { 0.0f, 0.0f, 0.0f };
-							t.scale = { 1.0f, 1.0f, 1.0f };
-							t.rotation = { 0.0f, 0.0f, 0.0f };
-
-							// 스키닝 메시 컴포넌트 등록
-							SkinnedMeshComponent& skinned = world.AddComponent<SkinnedMeshComponent>(e, result.meshAssetPath);
-							skinned.instanceAssetPath = result.instanceAssetPath;
-
-							// (임시) 본 행렬이 아직 없으므로, 1개짜리 항등 행렬 팔레트를 사용합니다.
-							//  - 나중에 FbxModel/FbxAnimation 연동 시 실제 본 팔레트로 교체됩니다.
-							static DirectX::XMFLOAT4X4 s_identityBone =
-								DirectX::XMFLOAT4X4(1, 0, 0, 0,
-									0, 1, 0, 0,
-									0, 0, 1, 0,
-									0, 0, 0, 1);
-							skinned.boneMatrices = &s_identityBone;
-							skinned.boneCount = 1;
-
-							// 첫 번째 머티리얼이 있으면 기본 머티리얼로 할당
-							// 원래 있는 경우 없는 경우 나눠서 있는 경우는 서브 메테리얼을 만들어야 하는데, 일단은 둘다 생기도록 함.
-							// TODO : 여기서 서브 메테리얼을 각각 다르게 설정할 수 있게 해야함 
-							if (!result.materialAssetPaths.empty())
-							{
-								DirectX::XMFLOAT3 defaultColor(0.7f, 0.7f, 0.7f);
-								MaterialComponent& mat = world.AddComponent<MaterialComponent>(e, defaultColor);
-								mat.assetPath = result.materialAssetPaths.front();
-								MaterialFile::Load(mat.assetPath, mat, &ResourceManager::Get());
-							}
-							else
-							{
-								//DirectX::XMFLOAT3 defaultColor(0.7f, 0.7f, 0.7f);
-								//MaterialComponent& mat = world.AddComponent<MaterialComponent>(e, defaultColor);
-								//mat.assetPath = "fbx has no material. default material";
-								//MaterialFile::Load(mat.assetPath, mat);
-							}
-
-							selectedEntity = e;
-							g_SceneDirty = true;
-						}
-					}
-				}
-			}
-
-			ImGui::Separator();
-
-			// 게임 빌드 버튼 (간단한 1차 버전)
-			if (ImGui::Button("Build"))
-			{
-				g_ShowBuildGameWindow = true;
-			}
-
-			ImGui::Separator();
-			// PVD 설정 버튼
-			if (ImGui::Button("PVD Settings"))
-			{
-				g_ShowPvdSettingsWindow = true;
-			}
-
-			ImGui::Separator();
-			ImGui::Text("DeltaTime: %.3f  FPS: %.1f", deltaTime, fps);
-
-			ImGui::Separator();
-			// 렌더링 시스템 선택 체크박스
-			ImGui::Checkbox("Show DebugDraw", &isDebugDraw);
-			if (ImGui::IsItemHovered())
-			{
-				ImGui::SetTooltip("체크: 디버그 라인 켜기\n해제: 디버그 라인 끄기");
-			}
-
-			ImGui::Separator();
-			// 렌더링 시스템 선택 체크박스
-			ImGui::Checkbox("Forward Rendering", &useForwardRendering);
-			if (ImGui::IsItemHovered())
-			{
-				ImGui::SetTooltip("체크: Forward Rendering\n해제: Deferred Rendering");
-			}
-
-			ImGui::EndMainMenuBar();
-		}
-
-		// === PVD Settings 창 ===
-		if (g_ShowPvdSettingsWindow)
-		{
-			static bool s_wasOpen = false;
-			bool isOpen = g_ShowPvdSettingsWindow;
-
-			if (ImGui::Begin("PVD Settings", &g_ShowPvdSettingsWindow))
-			{
-				ImGui::Text("PhysX Visual Debugger Settings");
-				ImGui::Separator();
-
-				// PVD 활성화 체크박스
-				ImGui::Checkbox("Enable PVD", &pvdEnabled);
-				if (ImGui::IsItemHovered())
-				{
-					ImGui::SetTooltip("Enable PhysX Visual Debugger.\n"
-						"Note: Requires restart to apply changes.\n"
-						"Make sure PVD is running on the target host/port.");
-				}
-
-				// PVD 설정 (비활성화 상태에서도 표시)
-				ImGui::BeginDisabled(!pvdEnabled);
-
-				// PVD Host 입력
-				static char pvdHostBuf[256] = {};
-				static bool s_hostBufInitialized = false;
-				if (!s_hostBufInitialized || !s_wasOpen)
-				{
-					strncpy_s(pvdHostBuf, pvdHost.c_str(), 255);
-					pvdHostBuf[255] = '\0';
-					s_hostBufInitialized = true;
-				}
-				ImGui::Text("Host:");
-				ImGui::SameLine();
-				if (ImGui::InputText("##PvdHost", pvdHostBuf, sizeof(pvdHostBuf)))
-				{
-					pvdHost = pvdHostBuf;
-				}
-				if (ImGui::IsItemHovered())
-				{
-					ImGui::SetTooltip("PVD server host (default: 127.0.0.1)");
-				}
-
-				// PVD Port 입력
-				ImGui::Text("Port:");
-				ImGui::SameLine();
-				if (ImGui::InputInt("##PvdPort", &pvdPort))
-				{
-					if (pvdPort < 1) pvdPort = 1;
-					if (pvdPort > 65535) pvdPort = 65535;
-				}
-				if (ImGui::IsItemHovered())
-				{
-					ImGui::SetTooltip("PVD server port (default: 5425)");
-				}
-
-				ImGui::EndDisabled();
-
-				ImGui::Separator();
-
-				// 상태 표시
-				ImGui::Text("Status:");
-				ImGui::SameLine();
-				if (pvdEnabled)
-				{
-					ImGui::TextColored(ImVec4(0.0f, 1.0f, 0.0f, 1.0f), "Enabled");
-					ImGui::Text("PVD will be enabled on next restart.");
-					ImGui::Text("Connection: %s:%d", pvdHost.c_str(), pvdPort);
-				}
-				else
-				{
-					ImGui::TextColored(ImVec4(0.7f, 0.7f, 0.7f, 1.0f), "Disabled");
-				}
-
-				ImGui::Separator();
-				ImGui::TextWrapped("Note: PVD settings are saved automatically when the engine shuts down.\n"
-					"Restart the engine to apply changes.");
-			}
-
-			// 창이 닫힐 때 설정 저장 (이전에 열려있었고 지금 닫힌 경우)
-			if (s_wasOpen && !g_ShowPvdSettingsWindow)
-			{
-				// 엔진 종료 시 자동 저장되므로 여기서는 선택적
-				// 필요시 여기서도 저장 가능
-			}
-			s_wasOpen = isOpen;
-
-			ImGui::End();
-		}
-
-		// === Build Game 창 (씬 선택 + 간단한 해상도 옵션) ===
-		if (g_ShowBuildGameWindow)
-		{
-			if (ImGui::Begin("Build Game", &g_ShowBuildGameWindow))
-			{
-				namespace fs = std::filesystem;
-
-				static int   s_Width = 1280;
-				static int   s_Height = 720;
-				static bool  s_ScanScenesOnce = true;
-				static std::vector<fs::path> s_ScenePaths;
-				static std::vector<bool>     s_SceneSelected;
-				static int   s_DefaultScene = -1;       // 기본으로 실행될 씬 인덱스
-				static char  s_ExportPath[260] = "../Build/Export"; // 배포용 출력 경로
-
-				ImGui::Text("Output Resolution");
-				ImGui::InputInt("Width (min : 320)", &s_Width);
-				ImGui::InputInt("Height (min : 240)", &s_Height);
-				if (s_Width < 320)  s_Width = 320;
-				if (s_Height < 240) s_Height = 240;
-
-				ImGui::Separator();
-				ImGui::Text("Scenes to Build");
-
-				if (s_ScanScenesOnce)
-				{
-					s_ScanScenesOnce = false;
-					s_ScenePaths.clear();
-					s_SceneSelected.clear();
-					s_DefaultScene = -1;
-
-					const fs::path assetsRoot = ResourceManager::Get().Resolve("Assets");
-					if (fs::exists(assetsRoot))
-					{
-						for (const auto& entry : fs::recursive_directory_iterator(assetsRoot))
-						{
-							if (!entry.is_regular_file())
-								continue;
-							if (entry.path().extension() != ".scene")
-								continue;
-
-							s_ScenePaths.push_back(entry.path());
-							s_SceneSelected.push_back(true);
-						}
-					}
-				}
-
-				// Refresh 버튼 추가
-				if (ImGui::Button("Refresh Scenes"))
-				{
-					s_ScanScenesOnce = true; // 다음 프레임에 다시 스캔
-				}
-				ImGui::SameLine();
-				ImGui::TextDisabled("(Click to rescan Assets folder)");
-
-				if (s_ScenePaths.empty())
-				{
-					ImGui::TextDisabled("No .scene files found under Assets.");
-				}
-				else
-				{
-					for (std::size_t i = 0; i < s_ScenePaths.size(); ++i)
-					{
-						bool selected = s_SceneSelected[i];
-						ImGui::Checkbox(s_ScenePaths[i].filename().string().c_str(), &selected);
-						s_SceneSelected[i] = selected;
-
-						ImGui::SameLine();
-						bool isDefault = (static_cast<int>(i) == s_DefaultScene);
-						std::string label = "Default##" + std::to_string(i);
-						if (ImGui::RadioButton(label.c_str(), isDefault))
-						{
-							s_DefaultScene = static_cast<int>(i);
-						}
-					}
-				}
-
-				ImGui::Separator();
-
-				// 배포용 출력 경로 입력 + 폴더 선택 버튼
-				ImGui::Text("Export Path (relative to project root or absolute)");
-				ImGui::InputText("##ExportPath", s_ExportPath, IM_ARRAYSIZE(s_ExportPath));
-
-				// Export Path 입력 필드에 드래그 앤 드롭 추가
-				if (ImGui::BeginDragDropTarget())
-				{
-					if (const ImGuiPayload* payload = ImGui::AcceptDragDropPayload("ASSET_FILE_PATH"))
-					{
-						const char* pathStr = static_cast<const char*>(payload->Data);
-						std::filesystem::path droppedPath(pathStr);
-						std::string ext = droppedPath.extension().string();
-						std::transform(ext.begin(), ext.end(), ext.begin(),
-							[](unsigned char c) { return static_cast<char>(std::tolower(c)); });
-
-						// 씬 파일(.scene)을 드래그하면 해당 씬 파일의 디렉토리 경로를 Export Path로 설정
-						if (ext == ".scene")
-						{
-							std::filesystem::path sceneDir = droppedPath.parent_path();
-							std::string dirStr = sceneDir.string();
-							strncpy_s(s_ExportPath, dirStr.c_str(), IM_ARRAYSIZE(s_ExportPath) - 1);
-							s_ExportPath[IM_ARRAYSIZE(s_ExportPath) - 1] = '\0';
-						}
-					}
-					ImGui::EndDragDropTarget();
-				}
-
-				ImGui::SameLine();
-				if (ImGui::Button("Browse..."))
-				{
-					BROWSEINFOW bi{};
-					bi.hwndOwner = m_hwnd;
-					bi.lpszTitle = L"Select export folder";
-					bi.ulFlags = BIF_RETURNONLYFSDIRS | BIF_USENEWUI;
-
-					PIDLIST_ABSOLUTE pidl = SHBrowseForFolderW(&bi);
-					if (pidl)
-					{
-						wchar_t folderW[MAX_PATH] = {};
-						if (SHGetPathFromIDListW(pidl, folderW))
-						{
-							std::filesystem::path p = folderW;
-							std::string utf8 = p.string();
-							// 선택한 경로를 그대로 ExportPath 로 사용 (필요하면 나중에 상대 경로로 변환 가능)
-							strncpy_s(s_ExportPath, utf8.c_str(), _TRUNCATE);
-						}
-						CoTaskMemFree(pidl);
-					}
-				}
-
-				// 빌드 진행 상황 표시
-				if (g_BuildInProgress.load())
-				{
-					ImGui::Text("Building AliceGame (Release)...");
-					float p = g_BuildProgress.load();
-					ImGui::ProgressBar(p, ImVec2(-1.0f, 0.0f));
-				}
-				else
-				{
-					long exitCode = g_BuildExitCode.load();
-					if (exitCode == 0)
-					{
-						ImGui::TextColored(ImVec4(0.3f, 0.9f, 0.3f, 1.0f), "Last build: Success");
-					}
-					else if (exitCode > 0)
-					{
-						ImGui::TextColored(ImVec4(0.9f, 0.3f, 0.3f, 1.0f), "Last build: Failed (code=%ld)", exitCode);
-					}
-				}
-
-				if (!g_BuildInProgress.load())
-				{
-					if (ImGui::Button("Build Game"))
-					{
-						// 1) 빌드 설정 파일 저장 (JSON)
-						wchar_t exePathW[MAX_PATH] = {};
-						GetModuleFileNameW(nullptr, exePathW, MAX_PATH);
-						fs::path exePath = exePathW;
-						fs::path exeDir = exePath.parent_path();
-						fs::path projectRoot = exeDir.parent_path().parent_path().parent_path(); // build/bin/Debug → 프로젝트 루트
-
-						fs::path buildDir = projectRoot / "Build";
-						std::error_code fec;
-						fs::create_directories(buildDir, fec);
-
-						fs::path cfgPath = buildDir / "BuildSettings.json";
-						{
-							std::ofstream ofs(cfgPath);
-							if (ofs.is_open())
-							{
-								nlohmann::json j;
-								j["width"] = s_Width;
-								j["height"] = s_Height;
-
-								std::vector<fs::path> includedScenes;
-								includedScenes.reserve(s_ScenePaths.size());
-								for (std::size_t i = 0; i < s_ScenePaths.size(); ++i)
-								{
-									if (i >= s_SceneSelected.size()) continue;
-									if (!s_SceneSelected[i]) continue;
-
-									fs::path relScene = fs::relative(s_ScenePaths[i], projectRoot);  // 프로젝트 루트 기준으로 상대 경로 (예: "Assets/Stage1/Stage1.scene")
-									includedScenes.push_back(relScene);
-								}
-
-								// 기본(default) 씬 선택
-								fs::path defaultScenePath;
-								bool validIndex =
-									s_DefaultScene >= 0 &&
-									static_cast<size_t>(s_DefaultScene) < s_ScenePaths.size() &&
-									s_DefaultScene < static_cast<int>(s_SceneSelected.size()) &&
-									s_SceneSelected[s_DefaultScene];
-
-								if (validIndex)
-								{
-									defaultScenePath = fs::relative(s_ScenePaths[s_DefaultScene], projectRoot); // 상대 경로로 가져오자. ../Assts를 Assets로 바꾸는 것
-								}
-								else if (!includedScenes.empty())
-								{
-									defaultScenePath = includedScenes.front();
-								}
-
-								if (!defaultScenePath.empty())
-								{
-									j["default"] = defaultScenePath.string();
-								}
-
-								std::vector<std::string> sceneStrings;
-								sceneStrings.reserve(includedScenes.size());
-								for (const auto& p : includedScenes)
-									sceneStrings.push_back(p.string());
-								j["scenes"] = sceneStrings;
-
-								ofs << j.dump(4);
-							}
-						}
-
-						ALICE_LOG_INFO("BuildSettings saved to \"%s\"", cfgPath.string().c_str());
-
-						// 2) 별도 스레드에서 CMake 빌드 + 리소스 복사 실행
-						g_BuildInProgress.store(true);
-						g_BuildProgress.store(0.0f);
-						g_BuildExitCode.store(-1);
-
-						// Export 경로 문자열은 스레드 시작 시점에 복사해 둡니다.
-						std::string exportPathStr = s_ExportPath;
-
-						const BuildGameTaskArgs args{ projectRoot, cfgPath, exportPathStr };
-						std::thread(&BuildGameTask::Run, args).detach();
-					}
-				}
-			}
-			ImGui::End();
-		}
-
-		// === Hierarchy ===
-		if (ImGui::Begin("Hierarchy"))
-		{
-			// "엔티티 목록" 텍스트에 드롭 타겟 추가 (부모 관계 해제)
-			Alice::ImGuiText(L"엔티티 목록");
-			if (ImGui::BeginDragDropTarget())
-			{
-				if (const ImGuiPayload* payload = ImGui::AcceptDragDropPayload("ENTITY_HIERARCHY"))
-				{
-					IM_ASSERT(payload->DataSize == sizeof(EntityId));
-					EntityId draggedId = *(const EntityId*)payload->Data;
-
-					if (draggedId != InvalidEntityId)
-					{
-						EntityId oldParent = world.GetParent(draggedId);
-						if (oldParent != InvalidEntityId)
-						{
-							// Transform 스냅샷 저장 (Undo용)
-							TransformComponent oldTransform;
-							if (auto* t = world.GetComponent<TransformComponent>(draggedId))
-							{
-								oldTransform = *t;
-							}
-
-							// keepWorld=true: 월드 위치 유지
-							world.SetParent(draggedId, InvalidEntityId, true);
-
-							// 성공 여부 확인 후에만 Undo 커맨드 추가
-							if (world.GetParent(draggedId) == InvalidEntityId)
-							{
-								// 새 Transform 스냅샷 저장
-								TransformComponent newTransform;
-								if (auto* t = world.GetComponent<TransformComponent>(draggedId))
-								{
-									newTransform = *t;
-								}
-								PushCommand(std::make_unique<SetParentCommand>(draggedId, oldParent, InvalidEntityId, oldTransform, newTransform));
-								g_SceneDirty = true;
-							}
-						}
-					}
-				}
-				ImGui::EndDragDropTarget();
-			}
-			
-			
-			ImGui::Separator();
-
-
-
-
-
-			EntityId entityToDelete = InvalidEntityId;
-			static EntityId s_renameTarget = InvalidEntityId;
-			static char s_renameBuf[128]{};
-			bool openRenamePopup = false;
-			static EntityId s_draggedEntity = InvalidEntityId;
-
-			// 재귀적으로 트리 노드를 그리는 람다 함수
-			std::function<void(EntityId)> DrawEntityNode = [&](EntityId entityId) {
-				if (entityId == InvalidEntityId)
-					return;
-
-				const bool isSelected = (selectedEntity == entityId);
-				const std::string name = world.GetEntityName(entityId);
-				const std::string label = !name.empty()
-					? name
-					: ("Entity " + std::to_string(static_cast<std::uint32_t>(entityId)));
-
-				ImGui::PushID((int)entityId);
-
-				// 자식들 가져오기
-				std::vector<EntityId> children = world.GetChildren(entityId);
-
-				// TreeNode 플래그
-				ImGuiTreeNodeFlags nodeFlags = ImGuiTreeNodeFlags_OpenOnArrow | ImGuiTreeNodeFlags_OpenOnDoubleClick;
-				if (isSelected)
-					nodeFlags |= ImGuiTreeNodeFlags_Selected;
-				if (children.empty())
-					nodeFlags |= ImGuiTreeNodeFlags_Leaf;
-
-			const bool isAliceUI = (world.GetComponent<UIWidgetComponent>(entityId) != nullptr);
-			if (isAliceUI)
-			{
-				ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(1.0f, 0.2f, 0.7f, 1.0f));
-			}
-
-			// 트리 노드 열기
-			bool nodeOpen = ImGui::TreeNodeEx(label.c_str(), nodeFlags);
-
-			if (isAliceUI)
-			{
-				ImGui::PopStyleColor();
-			}
-
-				// 선택 처리 (더블클릭으로만 인스펙터 변경 - 드래그앤드롭을 위해)
-				if (ImGui::IsItemHovered() && ImGui::IsMouseDoubleClicked(ImGuiMouseButton_Left))
-				{
-					selectedEntity = entityId;
-				}
-
-				// 드래그 시작
-				if (ImGui::BeginDragDropSource(ImGuiDragDropFlags_None))
-				{
-					ImGui::SetDragDropPayload("ENTITY_HIERARCHY", &entityId, sizeof(EntityId));
-					ImGui::TextUnformatted(label.c_str());
-					s_draggedEntity = entityId;
-					ImGui::EndDragDropSource();
-				}
-
-				// 드롭 대상 처리
-				if (ImGui::BeginDragDropTarget())
-				{
-					if (const ImGuiPayload* payload = ImGui::AcceptDragDropPayload("ENTITY_HIERARCHY"))
-					{
-						IM_ASSERT(payload->DataSize == sizeof(EntityId));
-						EntityId draggedId = *(const EntityId*)payload->Data;
-
-						if (draggedId != entityId && draggedId != InvalidEntityId)
-						{
-							// 기존 부모 가져오기
-							EntityId oldParent = world.GetParent(draggedId);
-
-							// Transform 스냅샷 저장 (Undo용)
-							TransformComponent oldTransform;
-							if (auto* t = world.GetComponent<TransformComponent>(draggedId))
-							{
-								oldTransform = *t;
-							}
-
-							// 새 부모 설정 (keepWorld=true: 월드 위치 유지)
-							world.SetParent(draggedId, entityId, true);
-
-							// 성공 여부 확인 후에만 Undo 커맨드 추가
-							if (world.GetParent(draggedId) == entityId)
-							{
-								// 새 Transform 스냅샷 저장
-								TransformComponent newTransform;
-								if (auto* t = world.GetComponent<TransformComponent>(draggedId))
-								{
-									newTransform = *t;
-								}
-								PushCommand(std::make_unique<SetParentCommand>(draggedId, oldParent, entityId, oldTransform, newTransform));
-								g_SceneDirty = true;
-							}
-						}
-					}
-					ImGui::EndDragDropTarget();
-				}
-
-				// 컨텍스트 메뉴
-				if (ImGui::BeginPopupContextItem())
-				{
-					if (ImGui::MenuItem("Change Name"))
-					{
-						s_renameTarget = entityId;
-						openRenamePopup = true;
-
-						const std::string cur = world.GetEntityName(entityId);
-						const std::string init = cur.empty()
-							? ("Entity " + std::to_string((std::uint32_t)entityId))
-							: cur;
-						std::memset(s_renameBuf, 0, sizeof(s_renameBuf));
-						strncpy_s(s_renameBuf, init.c_str(), sizeof(s_renameBuf) - 1);
-					}
-
-					if (ImGui::MenuItem("Delete"))
-					{
-						entityToDelete = entityId;
-					}
-
-					if (ImGui::MenuItem("Unparent"))
-					{
-						EntityId oldParent = world.GetParent(entityId);
-						if (oldParent != InvalidEntityId)
-						{
-							world.SetParent(entityId, InvalidEntityId);
-
-							// 성공 여부 확인 후에만 Undo 커맨드 추가
-							if (world.GetParent(entityId) == InvalidEntityId)
-							{
-								PushCommand(std::make_unique<SetParentCommand>(entityId, oldParent, InvalidEntityId));
-								g_SceneDirty = true;
-							}
-						}
-					}
-
-					// 현재 게임 오브젝트를 프리팹으로 저장하는 기능
-					if (ImGui::MenuItem("Save as Prefab"))
-					{
-						namespace fs = std::filesystem;
-						const fs::path prefabDir = ResourceManager::Get().Resolve("Assets/Prefabs");
-						if (!fs::exists(prefabDir))
-						{
-							fs::create_directories(prefabDir);
-						}
-
-						std::string baseName = "Entity_" + std::to_string(static_cast<std::uint32_t>(entityId)) + ".prefab";
-						fs::path prefabPath = prefabDir / baseName;
-
-						int index = 1;
-						while (fs::exists(prefabPath))
-						{
-							baseName = "Entity_" + std::to_string(static_cast<std::uint32_t>(entityId)) + "_" + std::to_string(index) + ".prefab";
-							prefabPath = prefabDir / baseName;
-							++index;
-						}
-
-						Prefab::SaveToFile(world, entityId, prefabPath);
-					}
-
-					ImGui::EndPopup();
-				}
-
-				// 자식 노드들 재귀적으로 그리기
-				if (nodeOpen)
-				{
-					// 자식들을 ID 순으로 정렬
-					std::sort(children.begin(), children.end());
-					for (EntityId child : children)
-					{
-						DrawEntityNode(child);
-					}
-				}
-
-				// TreeNodeEx를 호출했으면 항상 TreePop을 호출해야 함
-				if (nodeOpen)
-				{
-					ImGui::TreePop();
-				}
-
-				ImGui::PopID();
-			};
-
-			// 루트 엔티티들 가져오기
-			std::vector<EntityId> rootEntities = world.GetRootEntities();
-
-			// AliceUI 엔티티들도 Hierarchy에 포함 (TransformComponent 없는 경우 대비)
-			// 단, 부모가 있는 UI 위젯은 루트 목록에 다시 넣지 않는다.
-			std::set<EntityId> rootSet(rootEntities.begin(), rootEntities.end());
-			for (auto [id, widget] : world.GetComponents<UIWidgetComponent>())
-			{
-				if (world.GetParent(id) != InvalidEntityId)
-					continue;
-				if (rootSet.insert(id).second)
-					rootEntities.push_back(id);
-			}
-
-			if (rootEntities.empty())
-			{
-				Alice::ImGuiText(L"생성된 엔티티가 없습니다.");
-			}
-			else
-			{
-				// 루트 엔티티들을 ID 순으로 정렬
-				std::sort(rootEntities.begin(), rootEntities.end());
-
-				// 루트 노드들 그리기
-				for (EntityId rootId : rootEntities)
-				{
-					DrawEntityNode(rootId);
-				}
-			}
-
-			
-
-			// 빈 공간에 드롭하여 부모 관계 해제 또는 프리팹 인스턴스화
-			// 빈 공간을 감지하기 위해 InvisibleButton 사용
-			ImVec2 availSize = ImGui::GetContentRegionAvail();
-			if (availSize.y > 0)
-			{
-				ImGui::InvisibleButton("##HierarchyEmptySpace", availSize);
-				if (ImGui::BeginDragDropTarget())
-				{
-					// 엔티티 드래그앤드롭: 부모 관계 해제
-					if (const ImGuiPayload* payload = ImGui::AcceptDragDropPayload("ENTITY_HIERARCHY"))
-					{
-						IM_ASSERT(payload->DataSize == sizeof(EntityId));
-						EntityId draggedId = *(const EntityId*)payload->Data;
-
-						if (draggedId != InvalidEntityId)
-						{
-							EntityId oldParent = world.GetParent(draggedId);
-							if (oldParent != InvalidEntityId)
-							{
-								// Transform 스냅샷 저장 (Undo용)
-								TransformComponent oldTransform;
-								if (auto* t = world.GetComponent<TransformComponent>(draggedId))
-								{
-									oldTransform = *t;
-								}
-
-								// keepWorld=true: 월드 위치 유지
-								world.SetParent(draggedId, InvalidEntityId, true);
-
-								// 성공 여부 확인 후에만 Undo 커맨드 추가
-								if (world.GetParent(draggedId) == InvalidEntityId)
-								{
-									// 새 Transform 스냅샷 저장
-									TransformComponent newTransform;
-									if (auto* t = world.GetComponent<TransformComponent>(draggedId))
-									{
-										newTransform = *t;
-									}
-									PushCommand(std::make_unique<SetParentCommand>(draggedId, oldParent, InvalidEntityId, oldTransform, newTransform));
-									g_SceneDirty = true;
-								}
-							}
-						}
-					}
-					// 프리팹 파일 드래그앤드롭: 인스턴스화
-					else if (const ImGuiPayload* payload = ImGui::AcceptDragDropPayload("ASSET_FILE_PATH"))
-					{
-						const char* pathStr = static_cast<const char*>(payload->Data);
-						std::filesystem::path droppedPath(pathStr);
-						std::string ext = droppedPath.extension().string();
-						std::transform(ext.begin(), ext.end(), ext.begin(),
-							[](unsigned char c) { return static_cast<char>(std::tolower(c)); });
-
-						if (ext == ".prefab")
-						{
-							EntityId e = Alice::Prefab::InstantiateFromFile(world, droppedPath);
-							if (e != InvalidEntityId)
-							{
-								selectedEntity = e;
-								g_SceneDirty = true;
-							}
-						}
-					}
-					ImGui::EndDragDropTarget();
-				}
-			}
-
-			// Delete 키 입력 처리: Hierarchy 창이 포커스를 가지고 있고, 텍스트 입력 중이 아닐 때
-			const bool hierarchyTextInputActive = io.WantTextInput || ImGui::IsAnyItemActive();
-			if (selectedEntity != InvalidEntityId &&
-				ImGui::IsWindowFocused(ImGuiFocusedFlags_RootAndChildWindows) &&
-				!hierarchyTextInputActive &&
-				m_inputSystem &&
-				m_inputSystem->IsKeyPressed(Keyboard::Keys::Delete))
-			{
-				// 선택된 엔티티 삭제
-				const std::string entityName = world.GetEntityName(selectedEntity);
-				PushCommand(std::make_unique<DestroyEntityCommand>(selectedEntity, entityName, world));
-				world.DestroyEntity(selectedEntity);
-				selectedEntity = InvalidEntityId;
-				g_SceneDirty = true;
-			}
-
-
-			
-
-
-			if (openRenamePopup)
-				ImGui::OpenPopup("Change Name");
-
-			if (ImGui::BeginPopupModal("Change Name", nullptr, ImGuiWindowFlags_AlwaysAutoResize))
-			{
-				ImGui::InputText("Name", s_renameBuf, sizeof(s_renameBuf));
-				if (ImGui::Button("OK"))
-				{
-					const std::string oldName = world.GetEntityName(s_renameTarget);
-					const std::string newName = s_renameBuf;
-					world.SetEntityName(s_renameTarget, newName);
-					PushCommand(std::make_unique<SetEntityNameCommand>(s_renameTarget, oldName, newName));
-					g_SceneDirty = true;
-					ImGui::CloseCurrentPopup();
-				}
-				ImGui::SameLine();
-				if (ImGui::Button("Cancel"))
-				{
-					ImGui::CloseCurrentPopup();
-				}
-				ImGui::EndPopup();
-			}
-
-				// 루프가 끝난 뒤에 실제 삭제를 수행합니다. (반복 중 컨테이너 수정 방지)
-				if (entityToDelete != InvalidEntityId)
-				{
-					const std::string entityName = world.GetEntityName(entityToDelete);
-					PushCommand(std::make_unique<DestroyEntityCommand>(entityToDelete, entityName, world));
-					world.DestroyEntity(entityToDelete);
-					if (selectedEntity == entityToDelete)
-					{
-						selectedEntity = InvalidEntityId;
-					}
-					g_SceneDirty = true;
-				}
-
-			}
-
-		ImGui::End();
-
-
-			
-			// === Inspector ===
-			if (ImGui::Begin("Inspector")) {
-				// Delete 키 입력 처리: Inspector 창이 포커스를 가지고 있고, 텍스트 입력 중이 아닐 때
-				// isTextInputActive는 이미 DrawEditorUI 시작 부분에서 정의되어 있지만,
-				// Inspector 내부에서도 사용하므로 다시 체크 (Inspector가 포커스를 가진 경우를 위해)
-				const bool inspectorTextInputActive = io.WantTextInput || ImGui::IsAnyItemActive();
-
-			if (selectedEntity != InvalidEntityId &&
-				ImGui::IsWindowFocused(ImGuiFocusedFlags_RootAndChildWindows) &&
-				!inspectorTextInputActive &&
-				m_inputSystem &&
-				m_inputSystem->IsKeyPressed(Keyboard::Keys::Delete)) {
-				// 선택된 엔티티 삭제
-				const std::string entityName = world.GetEntityName(selectedEntity);
-				PushCommand(std::make_unique<DestroyEntityCommand>(selectedEntity, entityName, world));
-				world.DestroyEntity(selectedEntity);
-				selectedEntity = InvalidEntityId;
-				g_SceneDirty = true;
-			}
-
-			// 프리팹 드래그앤드롭: Inspector 창 전체에 드롭 타겟 추가
-			if (ImGui::BeginDragDropTarget())
-			{
-				if (const ImGuiPayload* payload = ImGui::AcceptDragDropPayload("ASSET_FILE_PATH"))
-				{
-					const char* pathStr = static_cast<const char*>(payload->Data);
-					std::filesystem::path droppedPath(pathStr);
-					std::string ext = droppedPath.extension().string();
-					std::transform(ext.begin(), ext.end(), ext.begin(),
-						[](unsigned char c) { return static_cast<char>(std::tolower(c)); });
-
-					if (ext == ".prefab")
-					{
-						if (selectedEntity != InvalidEntityId)
-						{
-							// 선택된 엔티티의 자식으로 추가
-							EntityId e = Alice::Prefab::InstantiateFromFile(world, droppedPath);
-							if (e != InvalidEntityId)
-							{
-								EntityId oldParent = world.GetParent(e);
-								world.SetParent(e, selectedEntity);
-
-								// 성공 여부 확인 후에만 Undo 커맨드 추가
-								if (world.GetParent(e) == selectedEntity)
-								{
-									PushCommand(std::make_unique<SetParentCommand>(e, oldParent, selectedEntity));
-									g_SceneDirty = true;
-								}
-
-								selectedEntity = e; // 새로 생성된 엔티티를 선택
-							}
-						}
-						else
-						{
-							// 선택된 엔티티가 없으면 루트에 추가
-							EntityId e = Alice::Prefab::InstantiateFromFile(world, droppedPath);
-							if (e != InvalidEntityId)
-							{
-								selectedEntity = e;
-								g_SceneDirty = true;
-							}
-						}
-					}
-				}
-				ImGui::EndDragDropTarget();
-			}
-
-				if (selectedEntity == InvalidEntityId) {
-					Alice::ImGuiText(L"선택된 엔티티가 없습니다.");
-				}
-				else {
-					// World 엔티티 Inspector 표시 (기존 로직)
-					// 엔티티 ID와 이름 표시 (한 번만 가져와서 재사용)
-					const std::string entityName = world.GetEntityName(selectedEntity);
-					if (!entityName.empty()) {
-						ImGui::Text("Entity %u - %s", static_cast<uint32_t>(selectedEntity), entityName.c_str());
-					}
-					else {
-						ImGui::Text("Entity %u", static_cast<uint32_t>(selectedEntity));
-					}
-					ImGui::Separator();
-
-				// 1. Transform
-				DrawInspectorTransform(world, selectedEntity);
-				ImGui::Separator();
-
-				// 1-1. Animation Status
-				DrawInspectorAnimationStatus(world, selectedEntity);
-				ImGui::Separator();
-
-				// 2. Scripts
-				ImGui::Text("Scripts");
-				DrawInspectorScripts(world, selectedEntity);
-				ImGui::Separator();
-
-				// 3. Material
-				DrawInspectorMaterial(world, selectedEntity);
-				ImGui::Separator();
-
-				// 3-2. Lights
-				DrawInspectorPointLight(world, selectedEntity);
-				DrawInspectorSpotLight(world, selectedEntity);
-				DrawInspectorRectLight(world, selectedEntity);
-				
-				// 3-3. Post Process Volume
-				DrawInspectorPostProcessVolume(world, selectedEntity);
-
-				// 3-3. Compute Effect
-				DrawInspectorComputeEffect(world, selectedEntity);
-				
-				// 3-4. Camera 컴포넌트들
-				DrawInspectorCameraSpringArm(world, selectedEntity);
-				DrawInspectorCameraLookAt(world, selectedEntity);
-				DrawInspectorCameraFollow(world, selectedEntity);
-				DrawInspectorCameraShake(world, selectedEntity);
-				DrawInspectorCameraInput(world, selectedEntity);
-				DrawInspectorCameraBlend(world, selectedEntity);
-				
-				// 4. Skinned Mesh / 소켓 프리뷰 (간단 뷰)
-				if (auto* skinned =
-					world.GetComponent<SkinnedMeshComponent>(selectedEntity)) {
-					ImGui::Separator();
-					ImGui::Text("Skinned Mesh: %s", skinned->meshAssetPath.c_str());
-
-					// 본 목록 미니 뷰 (이름 확인용)
-					if (m_skinnedRegistry) {
-						auto mesh = m_skinnedRegistry->Find(skinned->meshAssetPath);
-						if (mesh && mesh->sourceModel) {
-							const auto& bones = mesh->sourceModel->GetBoneNames();
-							if (ImGui::TreeNode("Bones")) {
-								for (size_t i = 0; i < bones.size(); ++i) {
-									ImGui::Text("%zu: %s", i, bones[i].c_str());
-								}
-								ImGui::TreePop();
-							}
-						}
-					}
-
-					// 메시 경로 필드에 드롭 타겟 추가
-					if (ImGui::BeginDragDropTarget())
-					{
-						if (const ImGuiPayload* payload = ImGui::AcceptDragDropPayload("ASSET_FILE_PATH"))
-						{
-							const char* pathStr = static_cast<const char*>(payload->Data);
-							std::filesystem::path droppedPath(pathStr);
-							std::string ext = droppedPath.extension().string();
-
-							// FBX 파일인지 확인
-							std::transform(ext.begin(), ext.end(), ext.begin(),
-								[](unsigned char c) { return static_cast<char>(std::tolower(c)); });
-							if (ext == ".fbx" || ext == ".fbxasset")
-							{
-								// 논리 경로로 변환
-								std::string logicalPath = droppedPath.string();
-								{
-									std::filesystem::path logical = ResourceManager::NormalizeResourcePathAbsoluteToLogical(droppedPath);
-									if (!logical.empty())
-									{
-										logicalPath = logical.string();
-									}
-								}
-								skinned->meshAssetPath = logicalPath;
-								g_SceneDirty = true;
-							}
-						}
-						ImGui::EndDragDropTarget();
-					}
-				}
-			}
-		}
-		ImGui::End();
-
-
-		// === Project ===
-		if (ImGui::Begin("Project"))
-		{
-			// 현재 씬 정보 및 저장 버튼
-			ImGui::Text("Current Scene:");
-			ImGui::SameLine();
-			if (g_HasCurrentScenePath)
-			{
-				std::string sceneName = g_CurrentScenePath.filename().string();
-				if (g_SceneDirty)
-				{
-					ImGui::TextColored(ImVec4(1.0f, 1.0f, 0.0f, 1.0f), "%s *", sceneName.c_str());
-				}
-				else
-				{
-					ImGui::Text("%s", sceneName.c_str());
-				}
-			}
-			else
-			{
-				ImGui::TextColored(ImVec4(0.7f, 0.7f, 0.7f, 1.0f), "Unsaved Scene");
-			}
-
-			ImGui::SameLine();
-			if (ImGui::Button("Save"))
-			{
-				SaveScene(world);
-			}
-			if (ImGui::IsItemHovered())
-			{
-				if (g_HasCurrentScenePath)
-				{
-					ImGui::SetTooltip("Save to: %s", g_CurrentScenePath.string().c_str());
-				}
-				else
-				{
-					ImGui::SetTooltip("Save scene (will use AutoSaved.scene if no path set)");
-				}
-			}
-
-			ImGui::Separator();
-
-			Alice::ImGuiText(L"Assets 폴더");
-			ImGui::Separator();
-
-			// Assets 폴더는 논리 경로로만 다루고, 실제 위치는 ResourceManager 가 해석합니다.
-			const std::filesystem::path assetsRoot = ResourceManager::Get().Resolve("Assets");
-			if (!std::filesystem::exists(assetsRoot))
-			{
-				// 폴더가 없다면 한 번만 생성해 둡니다.
-				std::filesystem::create_directories(assetsRoot);
-			}
-
-			DrawDirectoryNode(world, selectedEntity, assetsRoot);
-		}
-		ImGui::End();
-
-		// === Game ===
-		if (ImGui::Begin("Game"))
-		{
-			// 키보드 단축키로 Gizmo 모드 변경 (InputSystem 사용)
-			// 텍스트 입력 중이 아닐 때만 단축키 작동
-			if (m_inputSystem && !isTextInputActive)
-			{
-				using namespace DirectX;
-				if (m_inputSystem->IsKeyPressed(Keyboard::Keys::W)) gizmoOp = ImGuizmo::TRANSLATE;
-				if (m_inputSystem->IsKeyPressed(Keyboard::Keys::E)) gizmoOp = ImGuizmo::ROTATE;
-				if (m_inputSystem->IsKeyPressed(Keyboard::Keys::R)) gizmoOp = ImGuizmo::SCALE;
-				if (m_inputSystem->IsKeyPressed(Keyboard::Keys::X))
-				{
-					gizmoMode = (gizmoMode == ImGuizmo::LOCAL) ? ImGuizmo::WORLD : ImGuizmo::LOCAL;
-				}
-			}
-
-			// Gizmo Operation 선택 버튼
-			if (ImGui::RadioButton("Translate (W)", gizmoOp == ImGuizmo::TRANSLATE))
-				gizmoOp = ImGuizmo::TRANSLATE;
-			ImGui::SameLine();
-			if (ImGui::RadioButton("Rotate (E)", gizmoOp == ImGuizmo::ROTATE))
-				gizmoOp = ImGuizmo::ROTATE;
-			ImGui::SameLine();
-			if (ImGui::RadioButton("Scale (R)", gizmoOp == ImGuizmo::SCALE))
-				gizmoOp = ImGuizmo::SCALE;
-
-			// Gizmo Mode 선택 (Scale 모드에서는 World만 지원)
-			if (gizmoOp != ImGuizmo::SCALE)
-			{
-				ImGui::SameLine();
-				if (ImGui::RadioButton("Local (X)", gizmoMode == ImGuizmo::LOCAL))
-					gizmoMode = ImGuizmo::LOCAL;
-				ImGui::SameLine();
-				if (ImGui::RadioButton("World (X)", gizmoMode == ImGuizmo::WORLD))
-					gizmoMode = ImGuizmo::WORLD;
-			}
-			else
-			{
-				gizmoMode = ImGuizmo::LOCAL; // Scale은 항상 Local
-			}
-
-			// 게임 상태 표시 (한 줄, 색상 포함)
-			ImGui::SameLine();
-			ImGui::Text(" | ");
-			ImGui::SameLine();
-			ImVec4 stateColor = isPlaying ? ImVec4(0.0f, 1.0f, 0.0f, 1.0f) : ImVec4(1.0f, 0.0f, 0.0f, 1.0f);
-			ImGui::TextColored(stateColor, "%s", isPlaying ? "Playing" : "Stopped");
-
-			// 스냅 토글 버튼
-			ImGui::SameLine();
-			ImGui::Text(" | ");
-			ImGui::SameLine();
-			static bool showSnapSettings = false;
-			if (ImGui::SmallButton("Snap"))
-			{
-				showSnapSettings = !showSnapSettings;
-			}
-
-			// 스냅 설정 UI (토글이 켜져 있을 때만 표시)
-			if (showSnapSettings)
-			{
-				ImGui::Separator();
-				ImGui::Text("Snap Settings");
-
-				// Snap 모드 선택
-				ImGui::Text("Snap Mode:");
-				const char* snapModeItems[] = { "None", "Increment", "Object" };
-				int snapModeInt = static_cast<int>(snapMode);
-				if (ImGui::Combo("##SnapMode", &snapModeInt, snapModeItems, IM_ARRAYSIZE(snapModeItems)))
-				{
-					snapMode = static_cast<SnapMode>(snapModeInt);
-					gizmoSnap = (snapMode == SnapMode::Increment); // 레거시 호환성
-				}
-
-				// Snap 값 설정
-				if (snapMode == SnapMode::Increment)
-				{
-					ImGui::Indent();
-					switch (gizmoOp)
-					{
-					case ImGuizmo::TRANSLATE:
-						ImGui::DragFloat3("Snap Translation", &snapTranslation.x, 0.1f, 0.01f, 100.0f);
-						break;
-					case ImGuizmo::ROTATE:
-						ImGui::DragFloat("Snap Rotation (deg)", &snapRotation, 1.0f, 1.0f, 90.0f);
-						break;
-					case ImGuizmo::SCALE:
-						ImGui::DragFloat("Snap Scale", &snapScale, 0.1f, 0.1f, 10.0f);
-						break;
-					default:
-						break;
-					}
-					ImGui::Unindent();
-				}
-				else if (snapMode == SnapMode::Object)
-				{
-					ImGui::Indent();
-					ImGui::DragFloat("Snap Distance", &objectSnapDistance, 0.1f, 0.01f, 10.0f);
-
-					// 오브젝트 스냅 타입 선택 (Blender 스타일)
-					ImGui::Text("Snap To:");
-					const char* snapTypeItems[] = { "Center", "Vertex", "Edge", "Face" };
-					int snapTypeInt = static_cast<int>(objectSnapType);
-					if (ImGui::Combo("##SnapType", &snapTypeInt, snapTypeItems, IM_ARRAYSIZE(snapTypeItems)))
-					{
-						objectSnapType = static_cast<ObjectSnapType>(snapTypeInt);
-					}
-
-					// 현재 스냅 타입 설명
-					switch (objectSnapType)
-					{
-					case ObjectSnapType::Center:
-						ImGui::TextDisabled("Snap to object center");
-						break;
-					case ObjectSnapType::Vertex:
-						ImGui::TextDisabled("Snap to mesh vertices (if available)");
-						break;
-					case ObjectSnapType::Edge:
-						ImGui::TextDisabled("Snap to mesh edges (if available)");
-						break;
-					case ObjectSnapType::Face:
-						ImGui::TextDisabled("Snap to mesh face centers (if available)");
-						break;
-					}
-					ImGui::Unindent();
-				}
-				ImGui::Separator();
-			}
-
-			// 에디터 뷰포트는 톤매핑 완료(LDR) 텍스처를 표시해야 정상 색감이 나옵니다.
-			ID3D11ShaderResourceView* sceneSRV = nullptr;
-			float sceneWidth = 0.0f;
-			float sceneHeight = 0.0f;
-
-			if (useForwardRendering)
-			{
-				sceneSRV = forward.GetViewportSRV();
-				sceneWidth = static_cast<float>(forward.GetSceneWidth());
-				sceneHeight = static_cast<float>(forward.GetSceneHeight());
-			}
-			else
-			{
-				sceneSRV = deferred.GetViewportSRV();
-				sceneWidth = static_cast<float>(deferred.GetSceneWidth());
-				sceneHeight = static_cast<float>(deferred.GetSceneHeight());
-			}
-
-			if (sceneSRV)
-			{
-				ImVec2 avail = ImGui::GetContentRegionAvail();
-				ImVec2 size = avail;
-
-				if (sceneWidth > 0.0f && sceneHeight > 0.0f)
-				{
-					const float aspectScene = sceneWidth / sceneHeight;
-					const float aspectAvail = (avail.y > 0.0f) ? (avail.x / avail.y) : aspectScene;
-
-					if (aspectAvail > aspectScene)
-					{
-						size.x = avail.y * aspectScene;
-						size.y = avail.y;
-					}
-					else
-					{
-						size.x = avail.x;
-						size.y = avail.x / aspectScene;
-					}
-				}
-
-				// Image를 그린다
-				ImGui::Image(sceneSRV, size);
-
-				// 이미지가 화면에 그려진 사각형(픽셀) - Image 호출 직후에만 유효
-				ImVec2 imgMin = ImGui::GetItemRectMin();
-				ImVec2 imgMax = ImGui::GetItemRectMax();
-				ImVec2 imgSize = ImGui::GetItemRectSize();
-
-				if (m_aliceUIRenderer && m_hwnd && imgSize.x > 0.0f && imgSize.y > 0.0f && sceneWidth > 0.0f && sceneHeight > 0.0f)
-				{
-					POINT p = { static_cast<LONG>(imgMin.x), static_cast<LONG>(imgMin.y) };
-					::ScreenToClient(m_hwnd, &p);
-					m_aliceUIRenderer->SetScreenInputRect(
-						static_cast<float>(p.x),
-						static_cast<float>(p.y),
-						imgSize.x,
-						imgSize.y,
-						sceneWidth,
-						sceneHeight);
-
-					ImVec2 mousePos = ImGui::GetMousePos();
-					const float u = (mousePos.x - imgMin.x) / imgSize.x;
-					const float v = (mousePos.y - imgMin.y) / imgSize.y;
-					const float mx = u * sceneWidth;
-					const float my = v * sceneHeight;
-					m_aliceUIRenderer->SetScreenMouseOverride(mx, my);
-				}
-
-				// 프리팹 드래그앤드롭: 뷰포트 이미지 위에 드롭 타겟 추가
-				if (ImGui::BeginDragDropTarget())
-				{
-					if (const ImGuiPayload* payload = ImGui::AcceptDragDropPayload("ASSET_FILE_PATH"))
-					{
-						const char* pathStr = static_cast<const char*>(payload->Data);
-						std::filesystem::path droppedPath(pathStr);
-						std::string ext = droppedPath.extension().string();
-						std::transform(ext.begin(), ext.end(), ext.begin(),
-							[](unsigned char c) { return static_cast<char>(std::tolower(c)); });
-
-						if (ext == ".prefab")
-						{
-							// 마우스 위치를 이미지 내 상대 좌표(0~1)로 변환
-							ImVec2 mousePos = ImGui::GetMousePos();
-							float u = (mousePos.x - imgMin.x) / imgSize.x;
-							float v = (mousePos.y - imgMin.y) / imgSize.y;
-
-							// 이미지 영역 내에 있는지 확인
-							if (u >= 0.0f && u <= 1.0f && v >= 0.0f && v <= 1.0f)
-							{
-								// NDC 좌표로 변환
-								const float ndcX = 2.0f * u - 1.0f;
-								const float ndcY = 1.0f - 2.0f * v;
-
-								// 카메라에서 레이를 쏴서 월드 좌표 계산
-								using namespace DirectX;
-								XMMATRIX viewXM = camera.GetViewMatrix();
-								XMMATRIX projXM = camera.GetProjectionMatrix();
-								XMMATRIX invViewProj = XMMatrixInverse(nullptr, XMMatrixMultiply(viewXM, projXM));
-
-								// 카메라 앞 일정 거리(5미터)에 배치
-								XMVECTOR nearPoint = XMVectorSet(ndcX, ndcY, 0.0f, 1.0f);
-								XMVECTOR farPoint = XMVectorSet(ndcX, ndcY, 1.0f, 1.0f);
-
-								nearPoint = XMVector3TransformCoord(nearPoint, invViewProj);
-								farPoint = XMVector3TransformCoord(farPoint, invViewProj);
-
-								XMVECTOR dirWorld = XMVector3Normalize(XMVectorSubtract(farPoint, nearPoint));
-								XMFLOAT3 camPos = camera.GetPosition();
-								XMVECTOR originWorld = XMLoadFloat3(&camPos);
-
-								// 카메라 앞 5미터 위치에 배치
-								const float distance = 5.0f;
-								XMVECTOR spawnPos = XMVectorAdd(originWorld, XMVectorScale(dirWorld, distance));
-
-								XMFLOAT3 spawnPosition;
-								XMStoreFloat3(&spawnPosition, spawnPos);
-
-								// 프리팹 인스턴스화
-								EntityId e = Alice::Prefab::InstantiateFromFile(world, droppedPath);
-								if (e != InvalidEntityId)
-								{
-									if (auto* transform = world.GetComponent<TransformComponent>(e))
-									{
-										transform->position = spawnPosition;
-										world.MarkTransformDirty(e);
-									}
-									selectedEntity = e;
-									g_SceneDirty = true;
-								}
-							}
-						}
-					}
-					ImGui::EndDragDropTarget();
-				}
-
-				// ImGuizmo를 사용하여 선택된 엔티티 조작 (재생 중이 아닐 때만)
-				if (!isPlaying && selectedEntity != InvalidEntityId)
-				{
-					if (TransformComponent* transform = world.GetComponent<TransformComponent>(selectedEntity))
-					{
-						// Gizmo 조작 시작/종료 감지를 위한 static 변수
-						static bool wasUsingGizmo = false;
-						static EntityId lastGizmoEntity = InvalidEntityId;
-						static TransformCommand::TransformData gizmoStartTransform;
-
-						// View/Proj 행렬 준비 (XMFLOAT4X4로 변환)
-						XMMATRIX viewXM = camera.GetViewMatrix();
-						XMMATRIX projXM = camera.GetProjectionMatrix();
-
-						XMFLOAT4X4 viewMatrix, projMatrix;
-						XMStoreFloat4x4(&viewMatrix, viewXM);
-						XMStoreFloat4x4(&projMatrix, projXM);
-
-						// ComputeWorldMatrix()로 통일 (런타임과 동일한 규약)
-						using namespace DirectX;
-
-						// ComputeWorldMatrix()를 사용하여 월드 행렬 계산 (런타임과 동일)
-						XMMATRIX worldMatrixXM = world.ComputeWorldMatrix(selectedEntity);
-
-						// XMMATRIX를 float[16] 배열로 변환 (ImGuizmo 형식: row-major)
-						XMFLOAT4X4 worldMatrixFloat4x4;
-						XMStoreFloat4x4(&worldMatrixFloat4x4, worldMatrixXM);
-						float worldMatrix[16];
-						memcpy(worldMatrix, &worldMatrixFloat4x4, sizeof(worldMatrix));
-
-						// ImGuizmo에 직접 포인터 전달
-						const float* viewMat = reinterpret_cast<const float*>(viewMatrix.m);
-						const float* projMat = reinterpret_cast<const float*>(projMatrix.m);
-						float* objMat = worldMatrix;
-
-						// ImGuizmo 설정
-						ImGuizmo::SetOrthographic(false);
-						ImDrawList* drawList = ImGui::GetWindowDrawList();
-						ImGuizmo::SetDrawlist(drawList);
-						// SetRect는 실제 이미지가 그려진 사각형(픽셀)을 사용
-						// sceneWidth/sceneHeight는 GPU 렌더 타겟 해상도이므로 화면 픽셀과 다를 수 있음
-						ImGuizmo::SetRect(imgMin.x, imgMin.y, imgSize.x, imgSize.y);
-
-						// Snap 값 준비
-						float* snap = nullptr;
-						float snapValue[3] = { 0, 0, 0 }; // Snap 값을 받을 임시 배열
-						bool forceSnap = false;
-						// 텍스트 입력 중이 아닐 때만 Ctrl 스냅 작동
-						if (m_inputSystem && !isTextInputActive)
-						{
-							using namespace DirectX;
-							forceSnap = m_inputSystem->IsKeyDown(Keyboard::Keys::LeftControl) ||
-								m_inputSystem->IsKeyDown(Keyboard::Keys::RightControl);
-						}
-
-						// Increment 스냅 모드일 때만 ImGuizmo에 snap 전달
-						if (snapMode == SnapMode::Increment && (gizmoSnap || forceSnap))
-						{
-							if (gizmoOp == ImGuizmo::TRANSLATE)
-							{
-								snapValue[0] = snapTranslation.x;
-								snapValue[1] = snapTranslation.y;
-								snapValue[2] = snapTranslation.z;
-								snap = snapValue;
-							}
-							else if (gizmoOp == ImGuizmo::ROTATE)
-							{
-								snapValue[0] = snapRotation;
-								snap = snapValue;
-							}
-							else if (gizmoOp == ImGuizmo::SCALE)
-							{
-								snapValue[0] = snapScale;
-								snap = snapValue;
-							}
-						}
-
-						// Gizmo 조작 시작 감지: Manipulate 호출 전에 체크 (시작 시점 감지용)
-						if (lastGizmoEntity != selectedEntity)
-						{
-							// 다른 엔티티로 변경: 상태 리셋
-							wasUsingGizmo = false;
-							lastGizmoEntity = selectedEntity;
-						}
-
-						// Gizmo 조작 (worldMatrix 배열을 직접 넘겨주어 수정되게 함)
-						bool manipulated = ImGuizmo::Manipulate(viewMat, projMat, gizmoOp, gizmoMode, objMat, nullptr, snap);
-
-						// Gizmo 조작 시작/종료 감지: Manipulate 호출 후에 체크 (정확한 상태 반영)
-						bool isUsingGizmo = ImGuizmo::IsUsing();
-
-						// 조작 시작 감지: false → true
-						if (!wasUsingGizmo && isUsingGizmo && lastGizmoEntity == selectedEntity)
-						{
-							// 조작 시작: 현재 Transform을 old state로 저장
-							gizmoStartTransform.position = transform->position;
-							gizmoStartTransform.rotation = transform->rotation;
-							gizmoStartTransform.scale = transform->scale;
-							gizmoStartTransform.enabled = transform->enabled;
-							gizmoStartTransform.visible = transform->visible;
-						}
-
-						if (manipulated)
-						{
-							// 조작된 월드 행렬을 로컬 Transform으로 변환
-							// ImGuizmo가 반환한 worldMatrix는 조작된 월드 행렬이므로,
-							// 부모의 월드 행렬을 역으로 곱해서 로컬 Transform을 추출해야 함
-							using namespace DirectX;
-
-							// 조작된 월드 행렬을 XMMATRIX로 변환
-							XMMATRIX manipulatedWorldMatrix = XMLoadFloat4x4(reinterpret_cast<XMFLOAT4X4*>(worldMatrix));
-
-							// 부모의 월드 행렬 계산 (부모가 있는 경우)
-							XMMATRIX parentWorldMatrix = XMMatrixIdentity();
-							if (transform->parent != InvalidEntityId)
-							{
-								parentWorldMatrix = world.ComputeWorldMatrix(transform->parent);
-							}
-
-							// 부모의 월드 행렬을 역으로 곱해서 로컬 행렬 추출 (row-vector 컨벤션)
-							XMMATRIX parentWorldMatrixInv = XMMatrixInverse(nullptr, parentWorldMatrix);
-
-							// 오브젝트 스냅 모드: 다른 엔티티에 스냅 (스냅이 있으면 월드 행렬 수정)
-							XMMATRIX finalWorldMatrix = manipulatedWorldMatrix;
-							if (snapMode == SnapMode::Object && gizmoOp == ImGuizmo::TRANSLATE)
-							{
-								// 오브젝트 스냅 모드용 위치 (월드 공간)
-								XMFLOAT3 worldPosition;
-								XMStoreFloat3(&worldPosition, XMVector3TransformCoord(XMVectorZero(), manipulatedWorldMatrix));
-								XMVECTOR currentPos = XMLoadFloat3(&worldPosition);
-
-								float minDistance = objectSnapDistance;
-								XMFLOAT3 snappedPosition = worldPosition;
-								bool foundSnap = false;
-
-								// 모든 엔티티를 순회하며 가장 가까운 위치 찾기
-								for (auto&& [eid, otherTransform] : world.GetComponents<TransformComponent>())
-								{
-									if (eid == selectedEntity) continue; // 자기 자신은 제외
-
-									// 스냅 타입에 따라 타겟 위치 결정
-									// 초기값: 월드 위치로 설정 (폴백용)
-									XMMATRIX otherWorld = world.ComputeWorldMatrix(eid);
-									XMVECTOR bestSnapPos = otherWorld.r[3]; // 월드 위치 (translation 부분)
-									float bestSnapDist = objectSnapDistance;
-									bool hasMeshSnap = false;
-
-									switch (objectSnapType)
-									{
-									case ObjectSnapType::Center:
-										// 중심점 스냅: 월드 위치 계산 (이미 bestSnapPos에 설정됨)
-									{
-										XMVECTOR diff = currentPos - bestSnapPos;
-										float dist = XMVectorGetX(XMVector3Length(diff));
-										if (dist < bestSnapDist)
-										{
-											bestSnapDist = dist;
-											hasMeshSnap = true;
-										}
-									}
-									break;
-
-									case ObjectSnapType::Vertex:
-									case ObjectSnapType::Edge:
-									case ObjectSnapType::Face:
-									{
-										// 메시 데이터 접근하여 버텍스/엣지/면 스냅
-										if (auto* skinned = world.GetComponent<SkinnedMeshComponent>(eid))
-										{
-											if (m_skinnedRegistry && !skinned->meshAssetPath.empty())
-											{
-												auto mesh = m_skinnedRegistry->Find(skinned->meshAssetPath);
-												if (mesh && mesh->sourceModel)
-												{
-													// 어댑터 함수를 사용하여 월드 행렬 계산
-													XMMATRIX worldMatrix = world.ComputeWorldMatrix(eid);
-
-													const auto& vertices = mesh->sourceModel->GetCPUVertices();
-													const auto& indices = mesh->sourceModel->GetCPUIndices();
-
-													if (!vertices.empty())
-													{
-														if (objectSnapType == ObjectSnapType::Vertex)
-														{
-															// 버텍스 스냅: 모든 버텍스를 월드 공간으로 변환
-															for (const auto& vert : vertices)
-															{
-																XMVECTOR localPos = XMLoadFloat3(&vert.pos);
-																XMVECTOR worldPos = XMVector3TransformCoord(localPos, worldMatrix);
-
-																XMVECTOR diff = currentPos - worldPos;
-																float dist = XMVectorGetX(XMVector3Length(diff));
-
-																if (dist < bestSnapDist)
-																{
-																	bestSnapDist = dist;
-																	bestSnapPos = worldPos;
-																	hasMeshSnap = true;
-																}
-															}
-														}
-														else if (objectSnapType == ObjectSnapType::Edge && !indices.empty())
-														{
-															// 엣지 스냅: 인덱스를 사용해 엣지 중점 계산
-															// 삼각형 리스트 가정 (3개씩)
-															for (size_t i = 0; i < indices.size(); i += 3)
-															{
-																if (i + 2 >= indices.size()) break;
-
-																uint32_t i0 = indices[i];
-																uint32_t i1 = indices[i + 1];
-																uint32_t i2 = indices[i + 2];
-
-																if (i0 >= vertices.size() || i1 >= vertices.size() || i2 >= vertices.size())
-																	continue;
-
-																// 삼각형의 3개 엣지
-																XMVECTOR v0 = XMVector3TransformCoord(XMLoadFloat3(&vertices[i0].pos), worldMatrix);
-																XMVECTOR v1 = XMVector3TransformCoord(XMLoadFloat3(&vertices[i1].pos), worldMatrix);
-																XMVECTOR v2 = XMVector3TransformCoord(XMLoadFloat3(&vertices[i2].pos), worldMatrix);
-
-																// 각 엣지의 중점
-																XMVECTOR edgeMidpoints[3] = {
-																	(v0 + v1) * 0.5f,
-																	(v1 + v2) * 0.5f,
-																	(v2 + v0) * 0.5f
-																};
-
-																for (int e = 0; e < 3; ++e)
-																{
-																	XMVECTOR diff = currentPos - edgeMidpoints[e];
-																	float dist = XMVectorGetX(XMVector3Length(diff));
-
-																	if (dist < bestSnapDist)
-																	{
-																		bestSnapDist = dist;
-																		bestSnapPos = edgeMidpoints[e];
-																		hasMeshSnap = true;
-																	}
-																}
-															}
-														}
-														else if (objectSnapType == ObjectSnapType::Face && !indices.empty())
-														{
-															// 면 스냅: 삼각형 중심 계산
-															for (size_t i = 0; i < indices.size(); i += 3)
-															{
-																if (i + 2 >= indices.size()) break;
-
-																uint32_t i0 = indices[i];
-																uint32_t i1 = indices[i + 1];
-																uint32_t i2 = indices[i + 2];
-
-																if (i0 >= vertices.size() || i1 >= vertices.size() || i2 >= vertices.size())
-																	continue;
-
-																XMVECTOR v0 = XMVector3TransformCoord(XMLoadFloat3(&vertices[i0].pos), worldMatrix);
-																XMVECTOR v1 = XMVector3TransformCoord(XMLoadFloat3(&vertices[i1].pos), worldMatrix);
-																XMVECTOR v2 = XMVector3TransformCoord(XMLoadFloat3(&vertices[i2].pos), worldMatrix);
-
-																// 삼각형 중심 (3개 버텍스의 평균)
-																XMVECTOR faceCenter = (v0 + v1 + v2) / 3.0f;
-
-																XMVECTOR diff = currentPos - faceCenter;
-																float dist = XMVectorGetX(XMVector3Length(diff));
-
-																if (dist < bestSnapDist)
-																{
-																	bestSnapDist = dist;
-																	bestSnapPos = faceCenter;
-																	hasMeshSnap = true;
-																}
-															}
-														}
-													}
-												}
-											}
-										}
-
-										// 메시가 없으면 중심점으로 폴백 (이미 bestSnapPos에 월드 위치 설정됨)
-										if (!hasMeshSnap)
-										{
-											hasMeshSnap = true;
-										}
-										break;
-									}
-									}
-
-									if (hasMeshSnap)
-									{
-										XMVECTOR diff = currentPos - bestSnapPos;
-										float distance = XMVectorGetX(XMVector3Length(diff));
-
-										if (distance < minDistance)
-										{
-											minDistance = distance;
-											XMStoreFloat3(&snappedPosition, bestSnapPos);
-											foundSnap = true;
-										}
-									}
-								}
-
-								if (foundSnap)
-								{
-									// 스냅된 월드 위치를 최종 월드 행렬에 반영
-									finalWorldMatrix.r[3] = XMVectorSet(snappedPosition.x, snappedPosition.y, snappedPosition.z, 1.0f);
-								}
-							}
-
-							// 최종 월드 행렬을 로컬 행렬로 변환 (스냅 적용 여부와 관계없이 한 번만)
-							XMMATRIX localMatrix = finalWorldMatrix * parentWorldMatrixInv;
-
-							// 어댑터 함수를 사용하여 로컬 행렬을 TRS로 분해
-							XMFLOAT3 newPosition, newRotation, newScale;
-							if (DecomposeLocalMatrix(localMatrix, newPosition, newRotation, newScale))
-							{
-								if (gizmoOp == ImGuizmo::TRANSLATE)
-								{
-									transform->position = newPosition;
-								}
-								else if (gizmoOp == ImGuizmo::ROTATE)
-								{
-									transform->rotation = newRotation;  // (x=pitch, y=yaw, z=roll) 라디안
-								}
-								else if (gizmoOp == ImGuizmo::SCALE)
-								{
-									transform->scale = newScale;
-								}
-							}
-
-							// ImGuizmo로 Transform이 변경되었고 물리 컴포넌트가 있으면 텔레포트 자동 활성화
-							if (auto* rigidBody = world.GetComponent<Phy_RigidBodyComponent>(selectedEntity))
-							{
-								rigidBody->teleport = true;
-							}
-							if (auto* cct = world.GetComponent<Phy_CCTComponent>(selectedEntity))
-							{
-								cct->teleport = true;
-							}
-							world.MarkTransformDirty(selectedEntity);
-							g_SceneDirty = true;
-						}
-
-						// 조작 종료 감지: true → false
-						if (wasUsingGizmo && !isUsingGizmo && lastGizmoEntity == selectedEntity)
-						{
-							// 조작 종료: TransformCommand push
-							TransformCommand::TransformData newTransform;
-							newTransform.position = transform->position;
-							newTransform.rotation = transform->rotation;
-							newTransform.scale = transform->scale;
-							newTransform.enabled = transform->enabled;
-							newTransform.visible = transform->visible;
-
-							// Transform이 실제로 변경되었는지 확인 (float 비교는 epsilon 사용)
-							constexpr float kFloatEpsilon = 1e-6f;
-							auto FloatNotEqual = [](float a, float b) { return std::fabs(a - b) > kFloatEpsilon; };
-
-							bool hasChanged =
-								FloatNotEqual(gizmoStartTransform.position.x, newTransform.position.x) ||
-								FloatNotEqual(gizmoStartTransform.position.y, newTransform.position.y) ||
-								FloatNotEqual(gizmoStartTransform.position.z, newTransform.position.z) ||
-								FloatNotEqual(gizmoStartTransform.rotation.x, newTransform.rotation.x) ||
-								FloatNotEqual(gizmoStartTransform.rotation.y, newTransform.rotation.y) ||
-								FloatNotEqual(gizmoStartTransform.rotation.z, newTransform.rotation.z) ||
-								FloatNotEqual(gizmoStartTransform.scale.x, newTransform.scale.x) ||
-								FloatNotEqual(gizmoStartTransform.scale.y, newTransform.scale.y) ||
-								FloatNotEqual(gizmoStartTransform.scale.z, newTransform.scale.z) ||
-								(gizmoStartTransform.enabled != newTransform.enabled) ||
-								(gizmoStartTransform.visible != newTransform.visible);
-
-							if (hasChanged)
-							{
-								PushCommand(std::make_unique<TransformCommand>(
-									selectedEntity, gizmoStartTransform, newTransform));
-							}
-						}
-
-						// 상태 업데이트
-						wasUsingGizmo = isUsingGizmo;
-						lastGizmoEntity = selectedEntity;
-					}
-				}
-
-				// Delete 키 입력 처리: 뷰포트가 포커스를 가지고 있고, 텍스트 입력 중이 아닐 때
-				// 기즈모를 잡고 있어도 삭제 가능하도록 처리
-				if (selectedEntity != InvalidEntityId &&
-					ImGui::IsWindowFocused(ImGuiFocusedFlags_RootAndChildWindows) &&
-					!isTextInputActive &&
-					m_inputSystem &&
-					m_inputSystem->IsKeyPressed(Keyboard::Keys::Delete))
-				{
-					// 선택된 엔티티 삭제
-					const std::string entityName = world.GetEntityName(selectedEntity);
-					PushCommand(std::make_unique<DestroyEntityCommand>(selectedEntity, entityName, world));
-					world.DestroyEntity(selectedEntity);
-					selectedEntity = InvalidEntityId;
-					g_SceneDirty = true;
-				}
-
-				// 엔티티 선택 (Gizmo 위에 있지 않을 때만)
-				// 최종 빌드(Release)에서는 뷰포트 피커가 작동하지 않도록 함
-				if (ImGui::IsItemHovered() && ImGui::IsMouseClicked(ImGuiMouseButton_Left))
-				{
-					// Gizmo 위에 있지 않고 사용 중이 아닐 때만 선택 처리
-					if (!ImGuizmo::IsOver() && !ImGuizmo::IsUsing())
-					{
-						const ImVec2 mousePos = ImGui::GetIO().MousePos;
-
-						//피킹은 실제 이미지 사각형(imgMin, imgSize)을 기준으로 계산
-						// imagePos나 size를 사용하면 레터박스/패딩 때문에 위치가 어긋남
-						const float localX = mousePos.x - imgMin.x;
-						const float localY = mousePos.y - imgMin.y;
-
-							if (localX >= 0.0f && localX <= imgSize.x &&
-								localY >= 0.0f && localY <= imgSize.y)
-							{
-								// UV 좌표를 실제 이미지 크기 기준으로 계산
-								const float u = (imgSize.x > 0.0f) ? (localX / imgSize.x) : 0.0f;
-								const float v = (imgSize.y > 0.0f) ? (localY / imgSize.y) : 0.0f;
-								EntityId hit = picker.Pick(world, camera, m_skinnedRegistry, u, v);
-								selectedEntity = hit;
-							}
-						}
-					}
-			}
-			else
-			{
-				Alice::ImGuiText("씬 텍스처가 아직 준비되지 않았습니다.");
-			}
-		}
-		ImGui::End();
-
-		// === Camera / Animation (같은 영역, 탭) ===
-		if (ImGui::Begin("Camera"))
-		{
-			if (ImGui::BeginTabBar("##CameraTabs"))
-			{
-				if (ImGui::BeginTabItem("Camera"))
-				{
-					if (ImGui::BeginTable("CameraSplit", 2, ImGuiTableFlags_Resizable | ImGuiTableFlags_BordersInnerV))
-					{
-						// 왼쪽 카메라 정보 및 설정
-						ImGui::TableNextColumn();
-
-						Alice::ImGuiText(L"카메라 정보");
-						ImGui::Separator();
-
-						auto pos = camera.GetPosition();
-						ImGui::Text("Position : (%.2f, %.2f, %.2f)", pos.x, pos.y, pos.z);
-
-						ImGui::Separator();
-						Alice::ImGuiText(L"카메라 설정");
-						float fov = XMConvertToDegrees(camera.GetFovYRadians());
-						float nearP = camera.GetNearPlane();
-						float farP = camera.GetFarPlane();
-						bool changed = false;
-
-						changed |= ImGui::SliderFloat("FOV (deg)", &fov, 20.0f, 120.0f);
-						changed |= ImGui::DragFloat("Near Plane", &nearP, 0.01f, 0.01f, 10.0f, "%.3f");
-						changed |= ImGui::DragFloat("Far Plane", &farP, 1.0f, 10.0f, 5000.0f, "%.1f");
-						ImGui::SliderFloat("Move Speed", &cameraMoveSpeed, 0.1f, 50.0f, "%.2f");
-
-						if (changed)
-						{
-							nearP = (std::max)(nearP, 0.01f);
-							farP = (std::max)(farP, nearP + 0.1f);
-							camera.SetPerspective(XMConvertToRadians(fov), camera.GetAspectRatio(), nearP, farP);
-						}
-
-						// 오른쪽 버튼 및 액션
-						ImGui::TableNextColumn();
-						Alice::ImGuiText(L"기능");
-						ImGui::Separator();
-
-						if (ImGui::Button("Place Camera", { -FLT_MIN, 0.0f }))
-						{
-							EntityId e = world.CreateCamera();
-							auto* tc = world.GetComponent<TransformComponent>(e);
-							if (!tc) tc = &world.AddComponent<TransformComponent>(e);
-
-							tc->SetPosition(camera.GetPosition());
-							tc->SetRotation(camera.GetRotationQuat());
-							tc->SetScale(camera.GetScale());
-
-							if (auto* cc = world.GetComponent<CameraComponent>(e))
-							{
-								cc->SetFov(XMConvertToDegrees(camera.GetFovYRadians()));
-								cc->SetNear(camera.GetNearPlane());
-								cc->SetFar(camera.GetFarPlane());
-							}
-
-							// 커맨드 시스템에 등록하여 실행 취소(Ctrl+Z)가 가능하게 함
-							PushCommand(std::make_unique<CreateEntityCommand>(e, "Placed Camera"));
-
-							selectedEntity = e;
-							g_SceneDirty = true;
-						}
-
-						ImGui::EndTable();
-					}
-					ImGui::EndTabItem();
-				}
-
-				if (ImGui::BeginTabItem("Animation"))
-				{
-					if (selectedEntity == InvalidEntityId)
-					{
-						ImGui::TextUnformatted("No entity selected.");
-					}
-					else
-					{
-						SkinnedMeshComponent* skinned = world.GetComponent<SkinnedMeshComponent>(selectedEntity);
-						if (!skinned || skinned->meshAssetPath.empty())
-						{
-							ImGui::TextUnformatted("Selected entity has no SkinnedMesh.");
-						}
-						else
-						{
-							std::shared_ptr<SkinnedMeshGPU> mesh;
-							if (m_skinnedRegistry)
-								mesh = m_skinnedRegistry->Find(skinned->meshAssetPath);
-
-							ImGui::Text("Entity: %u", (unsigned)selectedEntity);
-							ImGui::Text("Mesh : %s", skinned->meshAssetPath.c_str());
-
-							if (!mesh || !mesh->sourceModel)
-							{
-								ImGui::Separator();
-								ImGui::TextUnformatted("Animation data is not ready (re-import FBX once).");
-								if (ImGui::Button("Re-import Skinned Meshes"))
-									EnsureSkinnedMeshesRegistered(world);
-							}
-							else
-							{
-								const auto& names = mesh->sourceModel->GetAnimationNames();
-								if (names.empty())
-								{
-									ImGui::TextUnformatted("This mesh has no animations.");
-								}
-								else
-								{
-									auto* anim = world.GetComponent<SkinnedAnimationComponent>(selectedEntity);
-									if (!anim) anim = &world.AddComponent<SkinnedAnimationComponent>(selectedEntity);
-
-									ImGui::Separator();
-									ImGui::Checkbox("Playing", &anim->playing);
-									ImGui::SliderFloat("Speed", &anim->speed, 0.0f, 3.0f, "%.2f");
-
-									int clip = anim->clipIndex;
-									if (clip < 0) clip = 0;
-									if (clip >= (int)names.size()) clip = (int)names.size() - 1;
-
-									if (ImGui::BeginCombo("Clip", names[(size_t)clip].c_str()))
-									{
-										for (int i = 0; i < (int)names.size(); ++i)
-										{
-											const bool sel = (i == clip);
-											if (ImGui::Selectable(names[(size_t)i].c_str(), sel))
-											{
-												clip = i;
-												anim->clipIndex = i;
-												anim->timeSec = 0.0;
-											}
-											if (sel) ImGui::SetItemDefaultFocus();
-										}
-										ImGui::EndCombo();
-									}
-									anim->clipIndex = clip;
-
-									const double dur = mesh->sourceModel->GetClipDurationSec(anim->clipIndex);
-									float timeSec = (float)anim->timeSec;
-									float durF = (dur > 0.0) ? (float)dur : 0.0f;
-
-									ImGui::BeginDisabled(durF <= 0.0f);
-									if (ImGui::SliderFloat("Time (sec)", &timeSec, 0.0f, durF, "%.3f"))
-										anim->timeSec = (double)timeSec;
-									ImGui::EndDisabled();
-
-									if (ImGui::Button("Stop"))
-									{
-										anim->playing = false;
-										anim->timeSec = 0.0;
-									}
-									ImGui::SameLine();
-									if (ImGui::Button("<<"))
-									{
-										anim->playing = false;
-										anim->timeSec = (std::max)(0.0, anim->timeSec - 0.1);
-									}
-									ImGui::SameLine();
-									if (ImGui::Button(">>"))
-									{
-										anim->playing = false;
-										anim->timeSec = anim->timeSec + 0.1;
-									}
-								}
-							}
-						}
-					}
-
-					ImGui::EndTabItem();
-				}
-
-				ImGui::EndTabBar();
-			}
-		}
-		ImGui::End();
-
-		// === Lighting ===
-		if (ImGui::Begin("Lighting"))
-		{
-			int mode = shadingMode;
-			if (ImGui::RadioButton("Lambert", mode == 0))   mode = 0;
-			ImGui::SameLine();
-			if (ImGui::RadioButton("Phong", mode == 1))     mode = 1;
-			ImGui::SameLine();
-			if (ImGui::RadioButton("Blinn-Phong", mode == 2)) mode = 2;
-			ImGui::SameLine();
-			if (ImGui::RadioButton("Toon", mode == 3))      mode = 3;
-			ImGui::SameLine();
-			if (ImGui::RadioButton("PBR", mode == 4))       mode = 4;
-			ImGui::SameLine();
-			if (ImGui::RadioButton("ToonPBR", mode == 5))   mode = 5;
-			ImGui::SameLine();
-			if (ImGui::RadioButton("ToonPBREditable", mode == 7)) mode = 7;
-			shadingMode = mode;
-
-			Alice::ImGuiCheckbox(L"Fill Light (보조광)", &useFillLight);
-
-			// Forward/Deferred 모드에 따라 조명 파라미터를 각 렌더러에 반영합니다.
-			//auto& lighting = useForwardRendering ? forward.GetLightingParameters() : deferred.GetLightingParameters();
-			//auto& lighting = forward.GetLightingParameters();
-			auto& lighting = deferred.GetLightingParameters();
-
-			// PBR 모드일 때 PBR 파라미터 표시
-			if (mode == 4 || mode == 5 || mode == 7)
-			{
-				ImGui::Separator();
-				ImGui::Text("PBR Material Parameters");
-				ImGui::ColorEdit3("Base Color", &lighting.baseColor.x);
-				ImGui::SliderFloat("Metalness", &lighting.metalness, 0.0f, 1.0f);
-				ImGui::SliderFloat("Roughness", &lighting.roughness, 0.0f, 1.0f);
-				ImGui::SliderFloat("Ambient Occlusion", &lighting.ambientOcclusion, 0.0f, 1.0f);
-				ImGui::Separator();
-			}
-			else
-			{
-				// 레거시 쉐이더 파라미터
-				ImGui::SliderFloat("Shininess", &lighting.shininess, 2.0f, 128.0f);
-				ImGui::ColorEdit3("Diffuse Color", &lighting.diffuseColor.x);
-				ImGui::ColorEdit3("Specular Color", &lighting.specularColor.x);
-			}
-
-			// 공통 조명 파라미터
-			Alice::ImGuiSliderFloat(L"Key Intensity (주광)",
-				&lighting.keyIntensity,
-				0.0f,
-				3.0f);
-			Alice::ImGuiSliderFloat(L"Fill Intensity (보조광)",
-				&lighting.fillIntensity,
-				0.0f,
-				3.0f);
-
-			Alice::ImGuiSliderFloat3(L"Key Direction (주광)",
-				&lighting.keyDirection.x,
-				-1.0f,
-				1.0f);
-			Alice::ImGuiSliderFloat3(L"Fill Direction (보조광)",
-				&lighting.fillDirection.x,
-				-1.0f,
-				1.0f);
-
-
-			// === Skybox ===
-			ImGui::Separator();
-			ImGui::TextUnformatted("Skybox");
-
-			static int  skyboxChoice = 3; // 0 Off, 1 Bridge, 2 Indoor, 3 Baker, 4 darkenv
-			static int  lastSkyboxChoice = -1;
-			static bool lastForward = false;
-
-			const char* skyboxItems[] = { "Off", "Bridge", "Indoor", "Baker", "darkenv"};
-
-			auto ApplySkybox = [&](auto& renderer)
-			{
-				if (skyboxChoice == 0)
-				{
-					renderer.SetSkyboxEnabled(false);
-					return;
-				}
-
-				renderer.SetSkyboxEnabled(true);
-				switch (skyboxChoice)
-				{
-				case 1: renderer.SetIblSet("Bridge", "bridge");       break;
-				case 2: renderer.SetIblSet("Indoor", "indoor");       break;
-				case 3: renderer.SetIblSet("Sample", "BakerSample");  break;
-				case 4: renderer.SetIblSet("darkenv", "darkenvDiffuseHDR");  break;
-				default: break;
-				}
-			};
-
-			auto EditBgIfOff = [&](auto& renderer)
-			{
-				if (skyboxChoice != 0) return;
-
-				DirectX::XMFLOAT4 bgColor = renderer.GetBackgroundColor();
-				if (ImGui::ColorEdit4("Background Color", &bgColor.x))
-					renderer.SetBackgroundColor(bgColor);
-			};
-
-			bool skyboxChanged = ImGui::Combo("Skybox Choice", &skyboxChoice, skyboxItems, IM_ARRAYSIZE(skyboxItems));
-			bool rendererChanged = (lastForward != useForwardRendering);
-
-			// 선택 변경 or 렌더러 토글 변경 시 반영 (초기 1회 포함)
-			if (skyboxChanged || rendererChanged || lastSkyboxChoice != skyboxChoice)
-			{
-				if (useForwardRendering) ApplySkybox(forward);
-				else                     ApplySkybox(deferred);
-
-				lastSkyboxChoice = skyboxChoice;
-				lastForward = useForwardRendering;
-			}
-
-			if (useForwardRendering) EditBgIfOff(forward);
-			else                     EditBgIfOff(deferred);
-
-
-			// === Post-Process (Exposure, Max HDR Nits) ===
-			ImGui::Separator();
-			ImGui::TextUnformatted("Post-Process");
-			ImGui::Separator();
-
-			float exposure = 0.0f;
-			float maxHDRNits = 1000.0f;
-			DirectX::XMFLOAT4 saturation = { 1.0f, 1.0f, 1.0f, 1.0f };
-			DirectX::XMFLOAT4 contrast = { 1.0f, 1.0f, 1.0f, 1.0f };
-			DirectX::XMFLOAT4 gamma = { 1.0f, 1.0f, 1.0f, 1.0f };
-			DirectX::XMFLOAT4 gain = { 1.0f, 1.0f, 1.0f, 1.0f };
-
-			auto DrawPostProcess = [&](auto& renderer)
-			{
-				// Exposure와 MaxHDRNits는 기존 함수로 가져오기
-				renderer.GetPostProcessParams(exposure, maxHDRNits);
-				// Color Grading은 Vector4로 가져오기
-				renderer.GetColorGrading(saturation, contrast, gamma, gain);
-
-				bool changed = false;
-
-				changed |= ImGui::SliderFloat("Exposure", &exposure, -3.0f, 3.0f, "%.2f");
-				if (ImGui::IsItemHovered())
-					ImGui::SetTooltip("Exposure 값: -3.0 (어두움) ~ 3.0 (밝음)\n0.0 = 1.0배 (기본값)");
-
-				changed |= ImGui::SliderFloat("Max HDR Nits", &maxHDRNits, 100.0f, 10000.0f, "%.0f nits");
-				if (ImGui::IsItemHovered())
-					ImGui::SetTooltip("HDR 모니터 최대 밝기 (nits)\n일반 모니터: 100-300 nits\nHDR 모니터: 1000-10000 nits");
-
-				ImGui::Separator();
-				ImGui::TextUnformatted("Color Grading (RGB 채널별 제어)");
-				
-				// ImGui::ColorEdit4는 float[4] 배열을 받지만, XMFLOAT4는 구조체이므로 배열로 변환 필요
-				// ImGuiFlags 설정: Alpha 슬라이더 숨김, Inputs 표시
-				ImGui::PushItemWidth(-1);
-
-				ImGui::Text("Saturation (RGB)");
-				changed |= ImGui::ColorEdit4("Saturation (RGB)", reinterpret_cast<float*>(&saturation),
-					ImGuiColorEditFlags_NoAlpha | ImGuiColorEditFlags_InputRGB | ImGuiColorEditFlags_Float);
-				if (ImGui::IsItemHovered())
-					ImGui::SetTooltip("채도 (R,G,B 채널별): 0.0 = 흑백, 1.0 = 원본, 2.0+ = 과포화\nW 채널은 항상 1.0으로 유지됩니다.");
-
-				ImGui::Text("Contrast (RGB)");
-				changed |= ImGui::ColorEdit4("Contrast (RGB)", reinterpret_cast<float*>(&contrast),
-					ImGuiColorEditFlags_NoAlpha | ImGuiColorEditFlags_InputRGB | ImGuiColorEditFlags_Float);
-				if (ImGui::IsItemHovered())
-					ImGui::SetTooltip("대비 (R,G,B 채널별): 0.0 = 회색, 1.0 = 원본, 2.0 = 고대비\nW 채널은 항상 1.0으로 유지됩니다.");
-				
-				ImGui::Text("Gamma (RGB)");
-				changed |= ImGui::ColorEdit4("Gamma (RGB)", reinterpret_cast<float*>(&gamma),
-					ImGuiColorEditFlags_NoAlpha | ImGuiColorEditFlags_InputRGB | ImGuiColorEditFlags_Float);
-				if (ImGui::IsItemHovered())
-					ImGui::SetTooltip("감마 보정 (R,G,B 채널별): 1.0 = 원본, <1.0 = 밝게, >1.0 = 어둡게\nW 채널은 항상 1.0으로 유지됩니다.");
-
-				ImGui::Text("Gain (RGB)");
-				changed |= ImGui::ColorEdit4("Gain (RGB)", reinterpret_cast<float*>(&gain),
-					ImGuiColorEditFlags_NoAlpha | ImGuiColorEditFlags_InputRGB | ImGuiColorEditFlags_Float);
-				if (ImGui::IsItemHovered())
-					ImGui::SetTooltip("Gain Multiply 스케일 (R,G,B 채널별): 0.0 = 검정, 1.0 = 원본, >1.0 = 밝게\nW 채널은 항상 1.0으로 유지됩니다.");
-				ImGui::PopItemWidth();
-
-				// W 채널은 항상 1.0으로 유지
-				saturation.w = 1.0f;
-				contrast.w = 1.0f;
-				gamma.w = 1.0f;
-				gain.w = 1.0f;
-
-				if (changed)
-				{
-					// Exposure와 MaxHDRNits는 기존 함수로 설정
-					renderer.SetPostProcessParams(exposure, maxHDRNits);
-					// Color Grading은 Vector4로 설정
-					renderer.ApplyColorGrading(saturation, contrast, gamma, gain);
-				}
-			};
-
-			if (useForwardRendering) DrawPostProcess(forward);
-			else                     DrawPostProcess(deferred);
-
-			// === Post Process Volume Reference Object ===
-			if (!useForwardRendering)
-			{
-				ImGui::Separator();
-				ImGui::TextUnformatted("Post Process Volume Reference");
-				ImGui::Separator();
-
-				static char refObjectNameBuf[256] = "";
-				std::string currentRefName = deferred.GetPPVReferenceObjectName();
-				if (strcmp(refObjectNameBuf, currentRefName.c_str()) != 0)
-				{
-					strncpy_s(refObjectNameBuf, currentRefName.c_str(), sizeof(refObjectNameBuf) - 1);
-					refObjectNameBuf[sizeof(refObjectNameBuf) - 1] = '\0';
-				}
-
-				if (ImGui::InputText("PPV Reference GameObject Name", refObjectNameBuf, sizeof(refObjectNameBuf)))
-				{
-					deferred.SetPPVReferenceObjectName(refObjectNameBuf);
-				}
-				if (ImGui::IsItemHovered())
-					ImGui::SetTooltip("PostProcessVolume 보간 기준이 될 GameObject 이름\n비어있으면 카메라 위치 사용");
-
-				// 현재 바인딩 상태 표시
-				if (!currentRefName.empty())
-				{
-					GameObject refObj = world.FindGameObject(currentRefName);
-					if (refObj.IsValid())
-					{
-						auto* transform = world.GetComponent<TransformComponent>(refObj.id());
-						if (transform && transform->enabled)
-						{
-							ImGui::TextColored(ImVec4(0.3f, 0.9f, 0.3f, 1.0f), 
-								"Bound to: %s (Position: %.2f, %.2f, %.2f)", 
-								currentRefName.c_str(), 
-								transform->position.x, transform->position.y, transform->position.z);
-						}
-						else
-						{
-							ImGui::TextColored(ImVec4(0.9f, 0.3f, 0.3f, 1.0f), 
-								"Bound to: %s (Transform not found or disabled)", currentRefName.c_str());
-						}
-					}
-					else
-					{
-						ImGui::TextColored(ImVec4(0.9f, 0.7f, 0.3f, 1.0f), 
-							"Object not found: %s (using camera position)", currentRefName.c_str());
-					}
-				}
-				else
-				{
-					ImGui::TextDisabled("Using camera position as reference");
-				}
-			}
-
-			// === Bloom (Deferred 전용) ===
-			if (!useForwardRendering)
-			{
-				ImGui::Separator();
-				ImGui::TextUnformatted("Bloom");
-				ImGui::Separator();
-
-				BloomSettings bloomSettings = deferred.GetBloomSettings();
-				bool bloomChanged = false;
-
-				if (ImGui::Checkbox("Enable Bloom", &bloomSettings.enabled))
-					bloomChanged = true;
-
-				if (ImGui::IsItemHovered())
-					ImGui::SetTooltip("Bloom 효과 활성화/비활성화");
-
-				if (bloomSettings.enabled)
-				{
-					if (ImGui::SliderFloat("Bloom Intensity", &bloomSettings.intensity, 0.0f, 5.0f, "%.2f"))
-						bloomChanged = true;
-					if (ImGui::IsItemHovered())
-						ImGui::SetTooltip("Bloom 합성 강도 (0.0 ~ 5.0)\n최종 합성 단계에서 적용되는 강도\n값이 클수록 더 밝게 합성됩니다");
-
-					if (ImGui::SliderFloat("Gaussian Intensity", &bloomSettings.gaussianIntensity, 0.0f, 5.0f, "%.2f"))
-						bloomChanged = true;
-					if (ImGui::IsItemHovered())
-						ImGui::SetTooltip("Gaussian 블러 강도 (0.0 ~ 5.0)\n블러 단계에서 적용되는 강도\n값이 클수록 블러 결과가 더 밝아집니다");
-
-					if (ImGui::SliderFloat("Threshold", &bloomSettings.threshold, 0.0f, 5.0f, "%.2f"))
-						bloomChanged = true;
-					if (ImGui::IsItemHovered())
-						ImGui::SetTooltip("밝기 추출 기준 (0.0 ~ 5.0)\n이 값보다 밝은 픽셀만 Bloom이 적용됩니다");
-
-					if (ImGui::SliderFloat("Knee", &bloomSettings.knee, 0.0f, 1.0f, "%.2f"))
-						bloomChanged = true;
-					if (ImGui::IsItemHovered())
-						ImGui::SetTooltip("Soft threshold (0.0 ~ 1.0)\nBloom 경계를 부드럽게 만드는 값");
-
-					if (ImGui::SliderFloat("Radius", &bloomSettings.radius, 0.0f, 20.0f, "%.1f"))
-						bloomChanged = true;
-					if (ImGui::IsItemHovered())
-						ImGui::SetTooltip("Blur 크기 (0.0 ~ 20.0)\n값이 클수록 더 넓게 퍼집니다");
-
-					const char* downsampleItems[] = {
-						"1x (원본)", "2x (1/2)", "4x (1/4)", "8x (1/8)",
-						"16x (1/16)", "32x (1/32)", "64x (1/64)"
-					};
-					const int downsampleValues[] = { 1, 2, 4, 8, 16, 32, 64 };
-
-					int downsampleIdx = 0;
-					for (int i = 0; i < 7; ++i)
-					{
-						if (bloomSettings.downsample == downsampleValues[i]) { downsampleIdx = i; break; }
-					}
-
-					if (ImGui::Combo("Downsample", &downsampleIdx, downsampleItems, IM_ARRAYSIZE(downsampleItems)))
-					{
-						bloomSettings.downsample = downsampleValues[downsampleIdx];
-						bloomChanged = true;
-					}
-					if (ImGui::IsItemHovered())
-						ImGui::SetTooltip("Bloom 다운샘플링 (1x ~ 64x)\n높을수록 성능↑ 품질↓");
-
-					if (ImGui::SliderFloat("Clamp", &bloomSettings.clamp, 1.0f, 20.0f, "%.1f"))
-						bloomChanged = true;
-					if (ImGui::IsItemHovered())
-						ImGui::SetTooltip("Bloom 값 상한 (1.0 ~ 20.0)\n과도한 Bloom을 제한합니다");
-				}
-
-				if (bloomChanged)
-					deferred.SetBloomSettings(bloomSettings);
-			}
-		}
-		ImGui::End();
-
-
-		// === Material Asset Editor (.mat 더블클릭 시) ===
-		if (g_MaterialEditorOpen)
-		{
-			if (ImGui::Begin("Material Asset Editor", &g_MaterialEditorOpen))
-			{
-				ImGui::Text("Asset: %s", g_MaterialEditorPath.string().c_str());
-				ImGui::Separator();
-
-				bool changed = false;
-				changed |= ImGui::ColorEdit3("Base Color", &g_MaterialEditorData.color.x);
-				changed |= ImGui::SliderFloat("Alpha", &g_MaterialEditorData.alpha, 0.0f, 1.0f);
-				changed |= ImGui::SliderFloat("Roughness", &g_MaterialEditorData.roughness, 0.0f, 1.0f);
-				changed |= ImGui::SliderFloat("Metalness", &g_MaterialEditorData.metalness, 0.0f, 1.0f);
-
-				ImGui::Separator();
-				ImGui::Text("Albedo Texture");
-				if (!g_MaterialEditorData.albedoTexturePath.empty())
-				{
-					ImGui::TextWrapped("%s", g_MaterialEditorData.albedoTexturePath.c_str());
-				}
-				else
-				{
-					ImGui::TextDisabled("None");
-				}
-				if (ImGui::Button("Browse Texture##Mat"))
-				{
-					wchar_t fileBuffer[MAX_PATH] = {};
-					OPENFILENAMEW ofn{};
-					ofn.lStructSize = sizeof(ofn);
-					ofn.hwndOwner = m_hwnd;
-					ofn.lpstrFilter = L"Image Files\0*.png;*.jpg;*.jpeg;*.tga;*.bmp;*.dds\0All Files\0*.*\0";
-					ofn.lpstrFile = fileBuffer;
-					ofn.nMaxFile = MAX_PATH;
-					ofn.Flags = OFN_EXPLORER | OFN_FILEMUSTEXIST | OFN_PATHMUSTEXIST | OFN_NOCHANGEDIR;
-
-					if (GetOpenFileNameW(&ofn))
-					{
-						std::filesystem::path absolutePath = fileBuffer;
-
-						// 절대 경로를 논리 경로로 변환
-						std::string logicalPath = absolutePath.string();
-						{
-							std::filesystem::path logical = ResourceManager::NormalizeResourcePathAbsoluteToLogical(absolutePath);
-							if (!logical.empty() && !logical.is_absolute())
-							{
-								logicalPath = logical.string();
-							}
-						}
-
-						g_MaterialEditorData.albedoTexturePath = logicalPath;
-						changed = true;
-
-						ALICE_LOG_INFO("[Editor] Material albedo set from MatEditor: \"%s\"\n",
-							g_MaterialEditorData.albedoTexturePath.c_str());
-					}
-				}
-
-				if (changed)
-				{
-					// 1) 에셋 파일에 저장
-					MaterialFile::Save(g_MaterialEditorPath, g_MaterialEditorData);
-
-					// 2) 이 에셋을 참조하는 모든 엔티티의 MaterialComponent 를 갱신
-					const std::string targetPath = g_MaterialEditorPath.string();
-					const auto& allMats = world.GetComponents<MaterialComponent>();
-					for (const auto& [id, matConst] : allMats)
-					{
-						MaterialComponent* mat = world.GetComponent<MaterialComponent>(id);
-						if (!mat) continue;
-						if (mat->assetPath == targetPath)
-						{
-							mat->color = g_MaterialEditorData.color;
-							mat->alpha = g_MaterialEditorData.alpha;
-							mat->roughness = g_MaterialEditorData.roughness;
-							mat->metalness = g_MaterialEditorData.metalness;
-						}
-					}
-
-					g_SceneDirty = true;
-				}
-			}
-			else
-			{
-				if (m_aliceUIRenderer)
-				{
-					m_aliceUIRenderer->ClearScreenInputRect();
-					m_aliceUIRenderer->ClearScreenMouseOverride();
-				}
-			}
-			ImGui::End();
-		}
-
-		// === UI Curve Asset Editor (.uicurve double-click) ===
-		if (g_UICurveEditorOpen)
-		{
-			if (ImGui::Begin("UI Curve Asset Editor", &g_UICurveEditorOpen))
-			{
-				ImGui::Text("Asset: %s", g_UICurveEditorPath.string().c_str());
-				ImGui::Separator();
-
-				bool changed = false;
-
-				ImVec2 graphSize = ImVec2(ImGui::GetContentRegionAvail().x, 180.0f);
-				ImVec2 graphPos = ImGui::GetCursorScreenPos();
-				ImGui::InvisibleButton("UICurveGraph", graphSize);
-				ImDrawList* drawList = ImGui::GetWindowDrawList();
-				ImVec2 graphMin = graphPos;
-				ImVec2 graphMax = ImVec2(graphPos.x + graphSize.x, graphPos.y + graphSize.y);
-				drawList->AddRect(graphMin, graphMax, IM_COL32(100, 100, 100, 255));
-
-				float tMin = 0.0f;
-				float tMax = 1.0f;
-				float vMin = 0.0f;
-				float vMax = 1.0f;
-				for (const auto& key : g_UICurveEditorData.keys)
-				{
-					tMin = std::min(tMin, key.time);
-					tMax = std::max(tMax, key.time);
-					vMin = std::min(vMin, key.value);
-					vMax = std::max(vMax, key.value);
-				}
-				const float tRange = std::max(0.0001f, tMax - tMin);
-				const float vRange = std::max(0.0001f, vMax - vMin);
-				auto ToScreen = [&](float t, float v)
-				{
-					const float x = (t - tMin) / tRange;
-					const float y = (v - vMin) / vRange;
-					return ImVec2(graphMin.x + x * graphSize.x, graphMax.y - y * graphSize.y);
-				};
-
-				for (int i = 1; i < 4; ++i)
-				{
-					const float tx = graphMin.x + (graphSize.x * i / 4.0f);
-					const float ty = graphMin.y + (graphSize.y * i / 4.0f);
-					drawList->AddLine(ImVec2(tx, graphMin.y), ImVec2(tx, graphMax.y), IM_COL32(60, 60, 60, 255));
-					drawList->AddLine(ImVec2(graphMin.x, ty), ImVec2(graphMax.x, ty), IM_COL32(60, 60, 60, 255));
-				}
-
-				if (!g_UICurveEditorData.keys.empty())
-				{
-					const int steps = 120;
-					ImVec2 prev = ToScreen(tMin, g_UICurveEditorData.Evaluate(tMin));
-					for (int i = 1; i < steps; ++i)
-					{
-						const float t = tMin + (tRange * (static_cast<float>(i) / (steps - 1)));
-						ImVec2 cur = ToScreen(t, g_UICurveEditorData.Evaluate(t));
-						drawList->AddLine(prev, cur, IM_COL32(120, 200, 255, 255), 2.0f);
-						prev = cur;
-					}
-				}
-
-				for (std::size_t i = 0; i < g_UICurveEditorData.keys.size(); ++i)
-				{
-					const auto& key = g_UICurveEditorData.keys[i];
-					ImVec2 p = ToScreen(key.time, key.value);
-					drawList->AddCircleFilled(p, 4.0f, IM_COL32(255, 200, 80, 255));
-					if (static_cast<int>(i) == g_UICurveEditorSelected)
-						drawList->AddCircle(p, 6.0f, IM_COL32(255, 255, 255, 200));
-				}
-
-				if (ImGui::IsItemHovered())
-				{
-					const ImVec2 mouse = ImGui::GetIO().MousePos;
-					if (ImGui::IsMouseClicked(ImGuiMouseButton_Left))
-					{
-						float bestDist = 999999.0f;
-						int bestIdx = -1;
-						for (std::size_t i = 0; i < g_UICurveEditorData.keys.size(); ++i)
-						{
-							ImVec2 p = ToScreen(g_UICurveEditorData.keys[i].time, g_UICurveEditorData.keys[i].value);
-							const float dx = mouse.x - p.x;
-							const float dy = mouse.y - p.y;
-							const float dist = dx * dx + dy * dy;
-							if (dist < bestDist)
-							{
-								bestDist = dist;
-								bestIdx = static_cast<int>(i);
-							}
-						}
-						if (bestIdx >= 0 && bestDist < 144.0f)
-							g_UICurveEditorSelected = bestIdx;
-					}
-					if (ImGui::IsMouseClicked(ImGuiMouseButton_Right))
-					{
-						float u = (mouse.x - graphMin.x) / std::max(1.0f, graphSize.x);
-						float v = 1.0f - (mouse.y - graphMin.y) / std::max(1.0f, graphSize.y);
-						u = std::clamp(u, 0.0f, 1.0f);
-						v = std::clamp(v, 0.0f, 1.0f);
-						UICurveKey key{};
-						key.time = tMin + u * tRange;
-						key.value = vMin + v * vRange;
-						key.interp = UICurveInterp::Cubic;
-						key.tangentMode = UICurveTangentMode::Auto;
-						g_UICurveEditorData.keys.push_back(key);
-						g_UICurveEditorSelected = static_cast<int>(g_UICurveEditorData.keys.size()) - 1;
-						changed = true;
-					}
-				}
-
-				ImGui::Separator();
-				if (ImGui::Button("Add Key"))
-				{
-					UICurveKey key{};
-					key.time = tMax;
-					key.value = 1.0f;
-					key.interp = UICurveInterp::Cubic;
-					key.tangentMode = UICurveTangentMode::Auto;
-					g_UICurveEditorData.keys.push_back(key);
-					g_UICurveEditorSelected = static_cast<int>(g_UICurveEditorData.keys.size()) - 1;
-					changed = true;
-				}
-				ImGui::SameLine();
-				if (ImGui::Button("Delete Key") && g_UICurveEditorSelected >= 0 && g_UICurveEditorSelected < static_cast<int>(g_UICurveEditorData.keys.size()))
-				{
-					g_UICurveEditorData.keys.erase(g_UICurveEditorData.keys.begin() + g_UICurveEditorSelected);
-					g_UICurveEditorSelected = -1;
-					changed = true;
-				}
-
-				const char* interpItems[] = { "Constant", "Linear", "Cubic" };
-				const char* tangentItems[] = { "Auto", "User", "Break" };
-
-				if (ImGui::BeginTable("UICurveKeys", 7, ImGuiTableFlags_Borders | ImGuiTableFlags_RowBg | ImGuiTableFlags_SizingStretchProp))
-				{
-					ImGui::TableSetupColumn("Idx");
-					ImGui::TableSetupColumn("Time");
-					ImGui::TableSetupColumn("Value");
-					ImGui::TableSetupColumn("Interp");
-					ImGui::TableSetupColumn("Tangent");
-					ImGui::TableSetupColumn("In");
-					ImGui::TableSetupColumn("Out");
-					ImGui::TableHeadersRow();
-					for (std::size_t i = 0; i < g_UICurveEditorData.keys.size(); ++i)
-					{
-						auto& key = g_UICurveEditorData.keys[i];
-						ImGui::TableNextRow();
-						ImGui::TableSetColumnIndex(0);
-						ImGui::PushID(static_cast<int>(i));
-						if (ImGui::Selectable(std::to_string(i).c_str(), g_UICurveEditorSelected == static_cast<int>(i)))
-							g_UICurveEditorSelected = static_cast<int>(i);
-						ImGui::TableSetColumnIndex(1);
-						changed |= ImGui::DragFloat("##time", &key.time, 0.01f);
-						ImGui::TableSetColumnIndex(2);
-						changed |= ImGui::DragFloat("##value", &key.value, 0.01f);
-						ImGui::TableSetColumnIndex(3);
-						int interpIdx = static_cast<int>(key.interp);
-						if (ImGui::Combo("##interp", &interpIdx, interpItems, IM_ARRAYSIZE(interpItems)))
-						{
-							key.interp = static_cast<UICurveInterp>(interpIdx);
-							changed = true;
-						}
-						ImGui::TableSetColumnIndex(4);
-						int tangentIdx = static_cast<int>(key.tangentMode);
-						if (ImGui::Combo("##tangent", &tangentIdx, tangentItems, IM_ARRAYSIZE(tangentItems)))
-						{
-							key.tangentMode = static_cast<UICurveTangentMode>(tangentIdx);
-							changed = true;
-						}
-						ImGui::TableSetColumnIndex(5);
-						changed |= ImGui::DragFloat("##in", &key.inTangent, 0.01f);
-						ImGui::TableSetColumnIndex(6);
-						changed |= ImGui::DragFloat("##out", &key.outTangent, 0.01f);
-						ImGui::PopID();
-					}
-					ImGui::EndTable();
-				}
-
-				if (changed)
-				{
-					g_UICurveEditorData.Sort();
-					g_UICurveEditorData.RecalcAutoTangents();
-					SaveUICurveAsset(g_UICurveEditorPath, g_UICurveEditorData);
-				}
-			}
-			ImGui::End();
-		}
-
-		// === 씬 변경사항 저장 확인 모달 ===
-		if (g_RequestSceneLoad)
-		{
-			// 현재 씬이 존재하고 변경사항이 있을 때만 확인 모달을 띄웁니다.
-			if (g_HasCurrentScenePath && g_SceneDirty)
-			{
-				ImGui::OpenPopup("SaveSceneBeforeLoad");
-			}
-			else
-			{
-				// 저장할 필요가 없으면 바로 로드
-				const std::filesystem::path loadAbs =
-					ResourceManager::Get().Resolve(g_NextScenePath);
-
-					if (isPlaying)
-					{
-						// 실행 중: 지연 처리
-						ALICE_LOG_INFO("[Editor] LoadSceneFileRequest (no-save, playing): \"%s\"\n",
-							g_NextScenePath.string().c_str());
-						if (sceneManager)
-						{
-							if (!sceneManager->LoadSceneFileRequest(loadAbs))
-							{
-								const std::string errorMsg = "씬 로드 요청 실패: " + g_NextScenePath.string() + "\n\n경로가 잘못되었거나 SceneManager가 초기화되지 않았습니다.";
-								ALICE_LOG_ERRORF("[Editor] Scene load request failed: %s", g_NextScenePath.string().c_str());
-								g_SceneLoadErrorMsg = errorMsg;
-								g_ShowSceneLoadError = true;
-							}
-							else
-							{
-								g_CurrentScenePath = g_NextScenePath;
-								g_HasCurrentScenePath = true;
-								g_SceneDirty = false;
-							}
-						}
-					}
-					else
-					{
-						// 실행 안 함: 즉시 로드
-						ALICE_LOG_INFO("[Editor] SceneFile::Load (no-save, not playing): \"%s\"\n",
-							g_NextScenePath.string().c_str());
-						const bool loadSuccess = SceneFile::LoadAuto(world, Alice::ResourceManager::Get(), g_NextScenePath);
-						
-						if (!loadSuccess)
-						{
-							const std::string errorMsg = "씬 로드 실패: " + g_NextScenePath.string() + "\n\n파일을 읽거나 역직렬화하는 중 오류가 발생했습니다.";
-							ALICE_LOG_ERRORF("[Editor] Scene load failed: %s", g_NextScenePath.string().c_str());
-							g_SceneLoadErrorMsg = errorMsg;
-							g_ShowSceneLoadError = true;
-						}
-						else
-						{
-							EnsureSkinnedMeshesRegistered(world);
-							selectedEntity = InvalidEntityId;
-							g_CurrentScenePath = g_NextScenePath;
-							g_HasCurrentScenePath = true;
-							g_SceneDirty = false;
-							ClearUndoStack(); // 씬 로드 시 Undo 스택 초기화
-						}
-					}
-				}
-				g_RequestSceneLoad = false;
-			}
-
-
-
-		// === 씬 로드 에러 모달 ===
-		if (g_ShowSceneLoadError)
-		{
-			ImGui::OpenPopup("SceneLoadError");
-			g_ShowSceneLoadError = false;
-		}
-
-		if (ImGui::BeginPopupModal("SceneLoadError", nullptr, ImGuiWindowFlags_AlwaysAutoResize))
-		{
-			ImGui::TextColored(ImVec4(1.0f, 0.3f, 0.3f, 1.0f), "씬 로드 실패");
-			ImGui::Separator();
-
-			// 에러 메시지 표시 (여러 줄 지원)
-			std::istringstream iss(g_SceneLoadErrorMsg);
-			std::string line;
-			while (std::getline(iss, line))
-			{
-				ImGui::TextWrapped("%s", line.c_str());
-			}
-
-			ImGui::Separator();
-			if (ImGui::Button("확인"))
-			{
-				g_SceneLoadErrorMsg.clear();
-				ImGui::CloseCurrentPopup();
-			}
-			ImGui::EndPopup();
-		}
-
-		if (ImGui::BeginPopupModal("SaveSceneBeforeLoad", nullptr, ImGuiWindowFlags_AlwaysAutoResize))
-		{
-			Alice::ImGuiText(L"현재 씬의 변경 내용을 저장하시겠습니까?");
-			ImGui::Separator();
-
-			if (ImGui::Button("Save"))
-			{
-				SaveScene(world);
-				// 씬 로드
-				const std::filesystem::path loadAbs =
-					ResourceManager::Get().Resolve(g_NextScenePath);
-
-					if (isPlaying)
-					{
-						// 실행 중: 지연 처리
-						if (sceneManager)
-						{
-							if (!sceneManager->LoadSceneFileRequest(loadAbs))
-							{
-								const std::string errorMsg = "씬 로드 요청 실패: " + g_NextScenePath.string() + "\n\n경로가 잘못되었거나 SceneManager가 초기화되지 않았습니다.";
-								ALICE_LOG_ERRORF("[Editor] Scene load request failed: %s", g_NextScenePath.string().c_str());
-								g_SceneLoadErrorMsg = errorMsg;
-								g_ShowSceneLoadError = true;
-								g_RequestSceneLoad = false;
-								ImGui::CloseCurrentPopup();
-								return;
-							}
-							g_CurrentScenePath = g_NextScenePath;
-							g_HasCurrentScenePath = true;
-							g_SceneDirty = false;
-						}
-						else
-						{
-							ALICE_LOG_ERRORF("[Editor] SceneManager is null, cannot load scene");
-						}
-					}
-					else
-					{
-						// 실행 안 함: 즉시 로드
-						const bool loadSuccess = SceneFile::LoadAuto(world, Alice::ResourceManager::Get(), g_NextScenePath);
-						
-						if (!loadSuccess)
-						{
-							const std::string errorMsg = "씬 로드 실패: " + g_NextScenePath.string() + "\n\n파일을 읽거나 역직렬화하는 중 오류가 발생했습니다.";
-							ALICE_LOG_ERRORF("[Editor] Scene load failed: %s", g_NextScenePath.string().c_str());
-							g_SceneLoadErrorMsg = errorMsg;
-							g_ShowSceneLoadError = true;
-							g_RequestSceneLoad = false;
-							ImGui::CloseCurrentPopup();
-							return;
-						}
-						EnsureSkinnedMeshesRegistered(world);
-						selectedEntity = InvalidEntityId;
-						g_CurrentScenePath = g_NextScenePath;
-						g_HasCurrentScenePath = true;
-						g_SceneDirty = false;
-					}
-					selectedEntity = InvalidEntityId;
-					g_RequestSceneLoad = false;
-					ImGui::CloseCurrentPopup();
-				}
-
-			ImGui::SameLine();
-			if (ImGui::Button("Don't Save"))
-			{
-				// 씬 로드
-				const std::filesystem::path loadAbs =
-					ResourceManager::Get().Resolve(g_NextScenePath);
-
-					if (isPlaying)
-					{
-						// 실행 중: 지연 처리
-						if (sceneManager)
-						{
-							ALICE_LOG_INFO("[Editor] LoadSceneFileRequest (dont-save, playing): \"%s\"\n",
-								g_NextScenePath.string().c_str());
-							if (!sceneManager->LoadSceneFileRequest(loadAbs))
-							{
-								const std::string errorMsg = "씬 로드 요청 실패: " + g_NextScenePath.string() + "\n\n경로가 잘못되었거나 SceneManager가 초기화되지 않았습니다.";
-								ALICE_LOG_ERRORF("[Editor] Scene load request failed: %s", g_NextScenePath.string().c_str());
-								g_SceneLoadErrorMsg = errorMsg;
-								g_ShowSceneLoadError = true;
-								g_RequestSceneLoad = false;
-								ImGui::CloseCurrentPopup();
-								return;
-							}
-							g_CurrentScenePath = g_NextScenePath;
-							g_HasCurrentScenePath = true;
-							g_SceneDirty = false;
-						}
-						else
-						{
-							ALICE_LOG_ERRORF("[Editor] SceneManager is null, cannot load scene");
-						}
-					}
-					else
-					{
-						// 실행 안 함: 즉시 로드
-						ALICE_LOG_INFO("[Editor] SceneFile::Load (dont-save, not playing): \"%s\"\n",
-							g_NextScenePath.string().c_str());
-						
-						const bool loadSuccess = SceneFile::LoadAuto(world, Alice::ResourceManager::Get(), g_NextScenePath);
-						
-						if (!loadSuccess)
-						{
-							const std::string errorMsg = "씬 로드 실패: " + g_NextScenePath.string() + "\n\n파일을 읽거나 역직렬화하는 중 오류가 발생했습니다.";
-							ALICE_LOG_ERRORF("[Editor] Scene load failed: %s", g_NextScenePath.string().c_str());
-							g_SceneLoadErrorMsg = errorMsg;
-							g_ShowSceneLoadError = true;
-							g_RequestSceneLoad = false;
-							ImGui::CloseCurrentPopup();
-							return;
-						}
-						EnsureSkinnedMeshesRegistered(world);
-						selectedEntity = InvalidEntityId;
-						g_CurrentScenePath = g_NextScenePath;
-						g_HasCurrentScenePath = true;
-						g_SceneDirty = false;
-					}
-					selectedEntity = InvalidEntityId;
-					g_RequestSceneLoad = false;
-					ImGui::CloseCurrentPopup();
-				}
-
-			ImGui::SameLine();
-			if (ImGui::Button("Cancel"))
-			{
-				// 아무것도 하지 않고 씬 로드를 취소합니다.
-				g_RequestSceneLoad = false;
-				ImGui::CloseCurrentPopup();
-			}
-
-			ImGui::EndPopup();
-		}
+		HandleSceneLoadFlow(world, sceneManager, isPlaying, selectedEntity);
 	}
 
 	void EditorCore::DrawInspectorTransform(World& world, const EntityId& _selectedEntity)
